% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{add_trajectory_incidence}
\alias{add_trajectory_incidence}
\alias{drop_trajectory_incidence}
\title{Add incidence to trajectories}
\usage{
add_trajectory_incidence(obj, states, suffix = "_inc")

drop_trajectory_incidence(obj)
}
\arguments{
\item{obj}{A \code{mcstate_trajectories} or \code{mcstate_pmcmc} object to
update}

\item{states}{A vector of cumulative states to compute incidence
for}

\item{suffix}{A string to append to the input states to create the
incidence variables}
}
\description{
Augment (or remove) trajectories from a \code{mcstate_trajectories} or
\code{mcstate_pmcmc} object.
}
