% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic}
\alias{basic}
\title{The basic sircovid model

The "basic" sircovid model. This is a dust model.}
\description{
The basic sircovid model

The "basic" sircovid model. This is a dust model.
}
\examples{
# Set up the basic model for England with default parameters and
# an initial seeding of early February:
p <- basic_parameters(sircovid_date("2020-02-07"), "england")
mod <- basic$new(p, 0, 10)

# Set the initial state and index as we would us for a run
# (without setting an initial state there is no seeding)
initial <- basic_initial(mod$info(), 10, p)
mod$set_state(initial$state, initial$step)
mod$set_index(basic_index(mod$info())$run)

# Run the model up to the end of march
step_end <- sircovid_date("2020-03-31") / p$dt

# The filtered state is returned at the end of the run
mod$run(step_end)

# More state can be retrieved using the "state" method
mod$state(1:6)
}
