% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_compare}
\alias{basic_compare}
\title{Compare observed and modelled data for the basic model}
\usage{
basic_compare(state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current day. This is
assumed to be filtered following \code{\link[=basic_index]{basic_index()}} so contains
rows corresponding to ICU and deaths.}

\item{observed}{Observed data. This will be a list with elements
\code{icu} (number of ICU beds occupied) and \code{deaths} (number of
deaths over this day).}

\item{pars}{A list of parameters, as created by
\code{\link[=basic_parameters]{basic_parameters()}}}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the basic model. This
conforms to the mcstate interface.
}
\examples{
state <- rbind(icu = 10:15, deaths_inc = 1:6)
observed <- list(icu = 13, deaths = 3)
pars <- basic_parameters(sircovid_date("2020-02-07"), "england")
basic_compare(state, observed, pars)
basic_compare(state * 5, observed, pars)
}
