% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_index}
\alias{basic_index}
\title{Index of basic model}
\usage{
basic_index(info)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an
initialised \link{basic} model}
}
\value{
A list with element \code{run}, indicating the locations of the
ICU and D compartments.
}
\description{
Index of "interesting" elements for the basic model. This function
conforms to the mcstate interface.
}
\examples{
p <- basic_parameters(sircovid_date("2020-02-07"), "england")
mod <- basic$new(p, 0, 10)
basic_index(mod$info())
}
