% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{basic_model}
\alias{basic_model}
\title{Basic model}
\usage{
basic_model(
  use_fitted_parameters = FALSE,
  progression_groups = list(E = 2, asympt = 1, mild = 1, ILI = 1, hosp = 2, ICU = 2,
    rec = 2),
  gammas = list(E = 1/(4.59/2), asympt = 1/2.09, mild = 1/2.09, ILI = 1/4, hosp = 2,
    ICU = 2/5, rec = 2/5)
)
}
\arguments{
\item{use_fitted_parameters}{Override progression_groups and gammas with fitted
parameters loaded by \code{read_fitted_parameters()}}

\item{progression_groups}{List of number of progression groups in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}

\item{gammas}{List of exponential distribution rates for time in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}
}
\description{
Create a basic model
}
