% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_parameters_observation}
\alias{basic_parameters_observation}
\title{Observation parameters for the basic model}
\usage{
basic_parameters_observation(exp_noise = 1e+06)
}
\arguments{
\item{exp_noise}{Rate of exponential noise - typically set to a
large value so that noise is small but non-zero. If set to \code{Inf}
then there is no noise in the observation process (not realistic
but useful for testing).}
}
\value{
A list of parameters
}
\description{
Paramters for the observation function for the basic model. Used
in \code{\link[=basic_compare]{basic_compare()}}
}
\examples{
basic_parameters_observation()
}
