% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes}
\alias{carehomes}
\title{The carehomes sircovid model

Our "current" sircovid model. This is a dust model.}
\description{
The carehomes sircovid model

Our "current" sircovid model. This is a dust model.
}
