% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{carehomes_check_severity}
\alias{carehomes_check_severity}
\title{Check severity parameters}
\usage{
carehomes_check_severity(pars)
}
\arguments{
\item{pars}{A list of probabilities from \link{carehomes_parameters}.}
}
\value{
Returns \code{pars} invisibly if all severity probabilities lie in
}
\description{
Checks severity probabilities are valid
}
