% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{carehomes_EpiEstim_Rt_trajectories}
\alias{carehomes_EpiEstim_Rt_trajectories}
\title{Compute Rt using EpiEstim for a set of trajectories}
\usage{
carehomes_EpiEstim_Rt_trajectories(
  step,
  incidence,
  p,
  sliding_window_ndays = 7,
  mean_prior = 1,
  sd_prior = 1,
  n_GT = 10000,
  n_R = 1000,
  save_all = TRUE
)
}
\arguments{
\item{step}{A vector of steps}

\item{incidence}{A matrix (n trajectories x n steps) of incidence counts}

\item{p}{A single \code{\link[=carehomes_parameters]{carehomes_parameters()}} object.}

\item{sliding_window_ndays}{An integer giving the length of the sliding
window on which Rt will be estimated}

\item{mean_prior}{The mean prior for Rt}

\item{sd_prior}{The standard deviation of the prior for Rt}

\item{n_GT}{An integer giving the number of generation times to be drawn to
construct the discrete distribution of the generation time}

\item{n_R}{An integer giving the number of Rt values to sample from for each
incidence trajectory. These will then be aggregated across all incidence
trajectories.}

\item{save_all}{A boolean determining whether to save all samples of Rt
estimated or only a summary}
}
\value{
A list with elements
\code{t_start} (vector of first days of the sliding windows over which Rt is
estimated),
\code{t_end} (vector of last days of the sliding windows over which Rt is
estimated),
\code{Rt} a matrix (only present if \code{save_all = TRUE}) containing for each
sliding window (each row in the matrix) a sample of n_R * nrow(inc) values
of Rt for that sliding window (columns of the matrix)
\code{Rt_summary} a matrix containing for each sliding window (each row in the
matrix) the 2.5\%, 50\%, 97.5\% quantiles and the mean of Rt for that sliding
window (columns of the matrix)
}
\description{
Compute "Rt" using EpiEtim for a set of simulated trajectories (e.g., the
result of the \verb{$iterate()} method of \link{carehomes}, \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}} or
\code{\link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict()}}. The trajectories should share
parameters.
}
