% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_forecast}
\alias{carehomes_forecast}
\title{Forecast the carehomes model}
\usage{
carehomes_forecast(
  samples,
  n_sample,
  burnin,
  forecast_days,
  incidence_states,
  prepend_trajectories = TRUE,
  random_sample = TRUE,
  thin = NULL
)
}
\arguments{
\item{samples}{Results of running \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}}}

\item{n_sample}{Number of samples to take. If you provide a value
of 0 then sampling is skipped (i.e., the entire set of
parameters in \code{samples} will be forecasted). Otherwise this is
passed to \link[mcstate:pmcmc_thin]{mcstate::pmcmc_sample}}

\item{burnin}{Optional integer number of iterations to discard as
"burn-in". If given then samples \code{1:burnin} will be
excluded from your results. Remember that the first sample
represents the starting point of the chain. It is an error if
this is not a positive integer or is greater than or equal to
the number of samples (i.e., there must be at least one sample
remaining after discarding burnin).}

\item{forecast_days}{The number of days to create a forecast for}

\item{incidence_states}{A character vector of states for which
incidnce should be computed (from cumulative compartments, such
as deaths). These will end up in the final trajectories object
with the sufix \verb{_inc} (e.g., \code{deaths} becomes \code{deaths_inc}).}

\item{prepend_trajectories}{Prepend trajectories from the particle
filter to the predictions created here.}

\item{random_sample}{Logical, whether to obtain the samples randomly
(using \link[mcstate:pmcmc_thin]{mcstate::pmcmc_sample}) or not (in which case the samples
are created via thinning with \link[mcstate:pmcmc_thin]{mcstate::pmcmc_thin})}

\item{thin}{Optional integer thinning factor, used if
\code{random_sample = FALSE}. After removal of burn-in, every \code{thin}'th
sample is retained}
}
\description{
Forecast from the carehomes model; this provides a wrapper around
\link[mcstate:pmcmc_thin]{mcstate::pmcmc_sample} (or \link[mcstate:pmcmc_thin]{mcstate::pmcmc_thin} if
\code{random_sample = FALSE}) and \link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict} that samples
the trajectories then creates forecast, setting the sircovid dates
and adding trajectories of incidence.
}
