% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_ifr_t.R
\name{carehomes_ifr_t}
\alias{carehomes_ifr_t}
\title{Compute "IFR_t"}
\usage{
carehomes_ifr_t(step, S, I_weighted, p)
}
\arguments{
\item{step}{A vector of steps that the model was run over}

\item{S}{A (n groups x n vaccine classes) x steps matrix of
"S" compartment counts}

\item{I_weighted}{A (n groups x n vaccine classes) x steps matrix of
"I_weighted" compartment counts}

\item{p}{A \code{\link[=carehomes_parameters]{carehomes_parameters()}} object}
}
\value{
A list with elements \code{step}, \code{date}, \code{IFR_t_general},
and `IFR_t_all``
}
\description{
Compute "IFR_t" for a single simulated trajectory and parameter set.
}
