% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_ifr_t.R
\name{carehomes_ifr_t_trajectories}
\alias{carehomes_ifr_t_trajectories}
\title{Compute IFR_t for a set of trajectories}
\usage{
carehomes_ifr_t_trajectories(
  step,
  S,
  I_weighted,
  pars,
  initial_step_from_parameters = TRUE,
  shared_parameters = NULL
)
}
\arguments{
\item{step}{A vector of steps}

\item{S}{A 3d ((n groups x n vaccine classes) x n trajectories
x n steps) array of "S" compartment counts}

\item{I_weighted}{A 3d ((n groups x n vaccine classes) x n trajectories
x n steps) array of "I_weighted" compartment counts}

\item{pars}{Either a single \code{\link[=carehomes_parameters]{carehomes_parameters()}} object
(shared parameters) or an unnamed list of
\code{\link[=carehomes_parameters]{carehomes_parameters()}} objects, the same length as \code{ncol(S)}.}

\item{initial_step_from_parameters}{If \code{TRUE}, then \code{step[[1]]} is
replaced by the value of \code{initial_step} from the parameters.
This is usually what you want.}

\item{shared_parameters}{Should \code{pars} be treated as a single
shared list? Leave as \code{NULL} to detect automatically, set to
\code{TRUE} or \code{FALSE} to force it to be interpreted one way or the
other which may give more easily interpretable error messages.}
}
\value{
As for \code{\link[=carehomes_ifr_t]{carehomes_ifr_t()}}, except that every element is a
matrix, not a vector.
}
\description{
Compute "IFR_t" for a set of simulated trajectories (e.g., the result
of \code{\link[dust:dust_iterate]{dust::dust_iterate()}}, \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}} or
\code{\link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict()}}. The trajectories may or may not share
parameters.
}
