% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_index}
\alias{carehomes_index}
\title{Index of carehomes model}
\usage{
carehomes_index(info)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an
initialised \link{basic} model}
}
\value{
A list with element \code{run}, indicating the locations of (in
order) (1) ICU, (2) general, (3) deaths in community, (4) deaths
in hospital, (5) total deaths, (6) cumulative confirmed
admissions,(7) cumulative confirmed new admissions, (8)
recovered pre seroconversion (15 - 64 year olds only), (9)
recovered seronegative and (10) recovered seropositive.
}
\description{
Index of "interesting" elements for the carehomes model. This function
conforms to the mcstate interface.
}
\examples{
p <- carehomes_parameters(sircovid_date("2020-02-07"), "england")
mod <- carehomes$new(p, 0, 10)
carehomes_index(mod$info())
}
