% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters}
\alias{carehomes_parameters}
\title{Parameters for the carehomes model}
\usage{
carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity_data = NULL
)
}
\arguments{
\item{start_date}{Thee start date, as a \code{\link[=sircovid_date]{sircovid_date()}} (i.e.,
the number of days into 2020)}

\item{region}{The region to run the model for. This will bee used
to get population data, which is currently fixed within the
package and is limited to "uk", the four constituent nations
("england", "wales", "scotland", "northern_ireland") and thee 7
NHS regions (e.g., "midlands"). These names are case
insensitive.}

\item{beta_date}{A vector of dates for changes in beta (the
contact rate parameter), or \code{NULL} if a single value is used for
all times (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{date}).}

\item{beta_value}{A vector of values for beta (the contact rate
parameter). If not given, and if \code{beta_date} is null then a
value of 0.08 will be used through the whole simulation,
otherwise if \code{beta_date} is \code{NULL} this must be a scalar. If
\code{beta_date} is given then \code{beta_date} and \code{beta_value} must have
the same length (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{value}).}

\item{severity_data}{A \code{data.frame} of severity data, or \code{NULL} to
use the default value within the package.  New severity data
comes from Bob Verity via the markovid package, and needs to be
carefully calibrated with the progression parameters.}
}
\value{
A list of inputs to the model, many of which are fixed and
represent data. These correspond largely to \code{user()} calls
within the odin code, though some are also used in processing
just before the model is run.
}
\description{
Parameters for the "\link{carehomes}" model.
}
\examples{
carehomes_parameters(sircovid_date("2020-02-01"), "uk")
}
