% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_observation}
\alias{carehomes_parameters_observation}
\title{Observation parameters for the carehomes model}
\usage{
carehomes_parameters_observation(exp_noise = 1e+06)
}
\arguments{
\item{exp_noise}{Rate of exponential noise - typically set to a
large value so that noise is small but non-zero. If set to \code{Inf}
then there is no noise in the observation process (not realistic
but useful for testing).}
}
\value{
A list of parameters
}
\description{
Parameters for the observation function for the carehomes model. Used
in \code{\link[=carehomes_compare]{carehomes_compare()}}
}
\examples{
carehomes_parameters_observation()
}
