% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_vaccination}
\alias{carehomes_parameters_vaccination}
\title{Carehomes vaccination parameters}
\usage{
carehomes_parameters_vaccination(rel_susc = 1)
}
\value{
A list of parameter values
}
\description{
Carehomes vaccination parameters.
}
