% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_particle_filter}
\alias{carehomes_particle_filter}
\title{Carehomes particle filter}
\usage{
carehomes_particle_filter(data, n_particles, n_threads = 1L, seed = NULL)
}
\arguments{
\item{data}{Data suitable for use with with the
\code{\link[mcstate:particle_filter]{mcstate::particle_filter}}, created by created by
\code{\link[mcstate:particle_filter_data]{mcstate::particle_filter_data()}}. We require columns "icu",
"general", "deaths_hosp", "deaths_comm", "deaths", "admitted",
"new", "npos_15_64", "ntot_15_64", though thse may be entirely
\code{NA} if no data are present.}

\item{n_particles}{Number of particles to use}

\item{n_threads}{Number of threads to use}

\item{seed}{Random seed to use}
}
\value{
A \code{\link[mcstate:particle_filter]{mcstate::particle_filter}} object
}
\description{
Construct a particle filter using the \code{\link{carehomes}} model. This is
a convenience function, ensuring that compatible functions are
used together.
}
