% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{carehomes_Rt_trajectories}
\alias{carehomes_Rt_trajectories}
\title{Compute Rt for a set of trajectories}
\usage{
carehomes_Rt_trajectories(
  step,
  S,
  pars,
  initial_step_from_parameters = TRUE,
  shared_parameters = NULL,
  type = NULL,
  interpolate_every = NULL,
  interpolate_critical = NULL,
  interpolate_min = NULL
)
}
\arguments{
\item{step}{A vector of steps}

\item{S}{A 3d (19 x n trajectories x n steps) array of "S"
compartment counts}

\item{pars}{Either a single \code{\link[=carehomes_parameters]{carehomes_parameters()}} object
(shared parameters) or an unnamed list of
\code{\link[=carehomes_parameters]{carehomes_parameters()}} objects, the same length as \code{ncol(S)}.}

\item{initial_step_from_parameters}{If \code{TRUE}, then \code{step[[1]]} is
replaced by the value of \code{initial_step} from the parameters.
This is usually what you want.}

\item{shared_parameters}{Should \code{pars} be treated as a single
shared list? Leave as \code{NULL} to detect automatically, set to
\code{TRUE} or \code{FALSE} to force it to be interpreted one way or the
other which may give more easily interpretable error messages.}

\item{type}{A character vector of possible Rt types to
compute. Can be any or all of \code{eff_Rt_all}, \code{eff_Rt_general},
\code{Rt_all} and \code{Rt_general}}

\item{interpolate_every}{Spacing to use between interpolated points}

\item{interpolate_critical}{Optional vector of critical step
indices to use when interpolating. Interpolation will be done in
blocks between these points, with each block starting on the
step given. So if you give a \code{interpolate_critical} of \code{c(20, 50)} then blocks \emph{start} on steps 20 and 50, i.e.: \verb{[1, 20)},
\verb{[20, 50)}, \verb{[50, end]}. Note that this is the index into the
\code{step} argument and not necessarily the same as the original
model step (being \code{dt * date}).}

\item{interpolate_min}{The minimum number of steps to include
within a block. If there are fewer points than this then all
points are used (i.e., no interpolation is done) or
\code{interpolate_every} is increased until at least this many points
were used. This can be used to specify a lower bound on the
error of small regions. If Rt is small it won't matter that
much. You do need to specify something though or interpolation
will not happen, and do not use less than 3 as we use spline
interpolation and that will not work with fewer than 3 points.}
}
\value{
As for \code{\link[=carehomes_Rt]{carehomes_Rt()}}, except that every element is a
matrix, not a vector.
}
\description{
Compute "Rt" for a set of simulated trajectories (e.g., the result
of \code{\link[dust:dust_iterate]{dust::dust_iterate()}}, \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}} or
\code{\link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict()}}. The trajectories may or may not share
parameters.
}
