% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_sircovid_model}
\alias{check_sircovid_model}
\title{Check a model type}
\usage{
check_sircovid_model(sircovid_model)
}
\arguments{
\item{sircovid_model}{String}
}
\description{
Check that a model type is valid (i.e. in \link{sircovid_models})
}
