% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{combine_rt_epiestim}
\alias{combine_rt_epiestim}
\title{Combine Rt estimates from EpiEstim}
\usage{
combine_rt_epiestim(rt, samples, q = NULL, rank = TRUE)
}
\arguments{
\item{rt}{A list of Rt calculations from
\code{\link[=lancelot_rt_trajectories_epiestim]{lancelot_rt_trajectories_epiestim()}} (though any Rt calculation that
confirms to this will work)}

\item{samples}{A list of samples from \link[mcstate:pmcmc]{mcstate::pmcmc}}

\item{q}{A vector of quantiles to return values for}

\item{rank}{A boolean deciding whether to rank trajectories by increasing
incidence or not before combining Rt estimates}
}
\value{
A list of Rt output in the same structure as the first
element of \code{rt}. Rt estimates will be aggregated across
regions (or whatever else you are aggregating on) based on the
parameters in \code{samples}.
}
\description{
Combine Rt estimates from EpiEstim across multiple runs.
}
