% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{compare_icu}
\alias{compare_icu}
\title{Compare model to ICU data}
\usage{
compare_icu(model, pars_obs, data)
}
\arguments{
\item{model}{An \code{odin_model} object}

\item{pars_obs}{Parameters for the observations}

\item{data}{The data to be compared against}
}
\description{
Compare the model to ICU data for use with the particle filter
}
