% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpu.R
\name{compile_gpu}
\alias{compile_gpu}
\title{Create GPU model}
\usage{
compile_gpu(model = "lancelot", ..., real_type = "float", gpu = TRUE)
}
\arguments{
\item{model}{The name of a sircovid model, either \code{lancelot} or
\code{basic}}

\item{...}{Additional arguments passed to \link[odin.dust:odin_dust]{odin.dust::odin_dust_}
and from there into either odin's options or the gpu options.}

\item{real_type}{The type to use for floating point numbers. The
default is "float" which differs from the CPU model but is
typically what you want on the GPU.}

\item{gpu}{The argument passed to \link[odin.dust:odin_dust]{odin.dust::odin_dust_} as
\code{gpu}. The default here is \code{TRUE}, but you may want to pass the
results of \link[dust:dust_cuda_options]{dust::dust_cuda_options} in order to control
compilation.}
}
\value{
A model generator, like \link{lancelot} that can run
on the GPU available to this computer.
}
\description{
Create a model for the GPU. This requires a working nvcc toolchain
and GPU device to work properly. Note that this will cache the
compilation within a session, so if you want to change GPU options
you will need to restart R. However, if you change the \emph{model}
(e.g., by changing the result of options like \code{rewrite_dims}) the
model will recompile. Compilation speed is slow enough that the
cachine will be obvious.
}
