% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{draw_one_GT_sample}
\alias{draw_one_GT_sample}
\title{Performs draws from the generation time distribution}
\usage{
draw_one_GT_sample(p, n = 1000)
}
\arguments{
\item{p}{A single \code{\link[=carehomes_parameters]{carehomes_parameters()}} object.}

\item{n}{The number of draws to perform}
}
\value{
a vector of \code{n} values of the generation time
}
\description{
Performs draws from the generation time distribution
}
