% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{drop_trajectory_predicted}
\alias{drop_trajectory_predicted}
\title{Drop predictions from trajectories}
\usage{
drop_trajectory_predicted(obj)
}
\arguments{
\item{obj}{A \code{mcstate_trajectories} or \code{mcstate_pmcmc} object to
update}
}
\description{
Drop predicted elements from a set of trajectories
}
