% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_beta_piecewise_linear}
\alias{generate_beta_piecewise_linear}
\title{Generate beta piecewise linear}
\usage{
generate_beta_piecewise_linear(
  beta_k,
  t_k,
  start_date = sircovid_date("2020-02-02"),
  dt
)
}
\arguments{
\item{beta_k}{Value of beta at changepoints}

\item{t_k}{Time of changepoints}

\item{start_date}{Start date for simulations}

\item{dt}{step size}
}
\description{
Create a piecewise linear time-varying beta for input into 
\code{create_parameters()}
}
