% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircovid.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(
  death_data_file,
  admissions_data_file,
  itu_colname = "number_of_confirmed_covid_19_patients_in_hdu_itu_at_0800_total",
  itu_ignore_dates = c("2020-03-15"),
  first_data_date = "2020-03-01",
  death_file_day_shift = 2,
  country_name = "United_Kingdom"
)
}
\arguments{
\item{death_data_file}{CSV file containing COVID-19 cases, from
the ECDC website}

\item{admissions_data_file}{Admission totals, from ncov-outputs}

\item{itu_colname}{Name of the column in the time series file
to be used for fitting ITU cases}

\item{itu_ignore_dates}{List of dates to discard from the
data due to error or otherwise}

\item{first_data_date}{Date to start data fit from}

\item{death_file_day_shift}{Number of days to subtract from
dateRep in case file by to match with date in 
\code{time_series_file}}

\item{country_name}{Country name in case reports}
}
\value{
Data for use with particle filter
}
\description{
Generate data for fitting the particle filter from reports of
deaths and admissions
}
