% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameter_rtm}
\alias{generate_parameter_rtm}
\title{Parameters for the rtm}
\usage{
generate_parameter_rtm(beta = 0.1, seed_SSP = 10, dt = 0.25)
}
\arguments{
\item{beta}{Beta}

\item{seed_SSP}{seed}

\item{dt}{dt}
}
\description{
Parameters for the rtm
}
