% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters_base}
\alias{generate_parameters_base}
\title{Create parameters base}
\usage{
generate_parameters_base(
  transmission_model,
  country,
  severity_data_file,
  beta,
  beta_times,
  importation_levels,
  importation_times,
  trans_profile,
  trans_increase,
  dt,
  use_polymod_pop,
  hospital_fitted
)
}
\arguments{
\item{transmission_model}{Model type. Only 'POLYMOD' currently supported}

\item{country}{Country name}

\item{severity_data_file}{Location of file with severity data}

\item{beta}{Beta, for each time step in \code{beta_times}}

\item{beta_times}{Dates \code{beta} changes, in format yyyy-mm-dd}

\item{importation_levels}{importation level, for each time step in \code{importation_times}}

\item{importation_times}{Dates importation changes, must use \code{sircovid_date}}

\item{trans_profile}{Proportion in each infectivity group}

\item{trans_increase}{Relative infectivity of each group}

\item{dt}{Time-step to run the model in days}

\item{use_polymod_pop}{Set to ignore \code{survey_pop_in}
and use the population from polymod when estimating the
transmission matrix}

\item{hospital_fitted}{If TRUE the parameters describing the progression
through the hospital compartments of the model are pulled from a set fitted
to a hospital model}
}
\value{
List of parameters for use with \code{sircovid}
}
\description{
General parameter creation for use with any model
}
