% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{hospital_model}
\alias{hospital_model}
\title{Hospital model}
\usage{
hospital_model(
  use_fitted_parameters = TRUE,
  progression_groups = list(E = 2, asympt = 1, mild = 1, ILI = 1, hosp_D = 2, hosp_R =
    2, ICU_D = 2, ICU_R = 2, triage = 2, stepdown = 2),
  gammas = list(E = 1/(4.59/2), asympt = 1/2.09, mild = 1/2.09, ILI = 1/4, hosp_D =
    2/5, hosp_R = 2/10, ICU_D = 2/5, ICU_R = 2/10, triage = 2, stepdown = 2/5)
)
}
\arguments{
\item{use_fitted_parameters}{Override progression_groups and gammas with fitted
parameters loaded by \code{read_fitted_parameters()}}

\item{progression_groups}{List of number of progression groups in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp_D', 'hosp_R', 'ICU_D', 'ICU_R', 'triage', 'stepdown'}

\item{gammas}{List of exponential distribution rates for time in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp_D', 'hosp_R', 'ICU_D', 'ICU_R', 'triage', 'stepdown'}
}
\description{
Create a hospital model
}
