% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{index_array}
\alias{index_array}
\title{Extract elements from an array when its dimensions are not known
in advance}
\usage{
index_array(x, dimension, value, drop = FALSE)
}
\arguments{
\item{x}{array}

\item{dimension}{which dimension to pull}

\item{value}{along the dimension \code{dimension}, which value to
extract.}

\item{drop}{logical. If TRUE the result is coerced to the lowest
possible dimension. Recommend setting this to FALSE}
}
\value{
the appropriate edge of the aray
}
\description{
Extract elements from an array when its dimensions are not known
in advance
}
\details{
As an axample, if the dimensions of x are 3, 4, 5 then
\code{index_array(x, dimension = 1, value = 2)} will return
x[2, , ]
}
