% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{inflate_state_strains}
\alias{inflate_state_strains}
\title{Inflate strains in model state}
\usage{
inflate_state_strains(state1, info1, info2)
}
\arguments{
\item{state1}{The state matrix with one strain}

\item{info1}{The model info with one strain}

\item{info2}{The model info with two strains}
}
\value{
An expanded model state with two strains
}
\description{
Inflate model state, so that additional strain compartments are
created but zerod. Use this to run the model for a while, then set
it up to work with new strains.
}
\author{
Richard Fitzjohn
}
