% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{inflate_state_vacc_classes}
\alias{inflate_state_vacc_classes}
\title{Inflate vacc classes in model state}
\usage{
inflate_state_vacc_classes(state1, info1, info2)
}
\arguments{
\item{state1}{The state matrix with X vacc classes}

\item{info1}{The model info with X vacc classes}

\item{info2}{The model info with Y > X vacc classes}
}
\value{
An expanded model state with Y > X vacc classes
}
\description{
Inflate model state, so that additional vacc class compartments are
created but zerod. Use this to run the model for a while, then set
it up to work with new vacc classes (e.g. booster doses).
}
