% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot}
\alias{lancelot}
\title{The lancelot sircovid model

Our "current" sircovid model. This is a dust model.}
\description{
The lancelot sircovid model

Our "current" sircovid model. This is a dust model.
}
