% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_check_data}
\alias{lancelot_check_data}
\title{Check data for particle filter}
\usage{
lancelot_check_data(data)
}
\arguments{
\item{data}{A data.frame of data}
}
\value{
Invisibly, a data frame, identical to \code{data}
}
\description{
Check that data for the particle filter has required columns and
constraints among columns.  This function does not alter the data
at present (though it does return it invisibly).
}
