% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_check_severity}
\alias{lancelot_check_severity}
\title{Check severity parameters}
\usage{
lancelot_check_severity(pars)
}
\arguments{
\item{pars}{A list of probabilities from \link{lancelot_parameters}.}
}
\value{
Returns \code{pars} invisibly if all severity probabilities lie in
}
\description{
Checks severity probabilities are valid
}
