% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot_ifr.R
\name{lancelot_ifr_excl_immunity}
\alias{lancelot_ifr_excl_immunity}
\title{Compute "IFR" excluding immunity}
\usage{
lancelot_ifr_excl_immunity(step, pars)
}
\arguments{
\item{step}{A vector of time steps to calculate IFR at}

\item{pars}{An unnamed list of \code{\link[=lancelot_parameters]{lancelot_parameters()}} objects}
}
\description{
Compute "IFR" excluding immunity.
}
