% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_index}
\alias{lancelot_index}
\title{Index of lancelot model}
\usage{
lancelot_index(
  info,
  rt = FALSE,
  cum_admit = FALSE,
  diagnoses_admitted = FALSE,
  cum_infections_disag = FALSE,
  cum_n_vaccinated = FALSE
)
}
\arguments{
\item{info}{The result of running the \verb{$info()} method on an
initialised \link{basic} model}

\item{rt}{logical parameter output trajectories required for
calculating Rt (default = FALSE)}

\item{cum_admit}{logical parameter whether to output cumulative
admissions by age (default = FALSE)}

\item{diagnoses_admitted}{logical parameter whether to output
cumulative combined confirmed admissions and inpatient
diagnoses by age and vaccine class (default = FALSE)}

\item{cum_infections_disag}{logical parameter whether to output
cumulative infections by age and vaccine class (default = FALSE)}

\item{cum_n_vaccinated}{logical parameter whether to output
cumulative number vaccinated by age and vaccine class
(default = FALSE)}
}
\value{
A list with element \code{run}, indicating the locations of (in
order) (1) ICU, (2) general, (3) deaths in community, (4) deaths
in hospital, (5) total deaths, (6) cumulative confirmed
admissions,(7) cumulative confirmed new admissions, (8)
probability of a positive sero test, (9) probability of a positive
pillar 2 test, and with element \code{state} containing the same values
followed by 17 S compartments (one per age group, then one for
carehome workers and carehome residents respectively) and 17
"cumulative admission" compartments.
}
\description{
Index of "interesting" elements for the lancelot model. This function
conforms to the mcstate interface.
}
\examples{
p <- lancelot_parameters(sircovid_date("2020-02-07"), "england")
mod <- lancelot$new(p, 0, 10)
lancelot_index(mod$info())
}
