% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_parameters}
\alias{lancelot_parameters}
\title{Parameters for the lancelot model}
\usage{
lancelot_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  beta_type = "piecewise-linear",
  population = NULL,
  carehome_beds = NULL,
  severity = NULL,
  progression = NULL,
  observation = NULL,
  sens_and_spec = NULL,
  initial_seed_size = 30,
  initial_seed_pattern = 1,
  eps = 0.1,
  m_CHW = 4e-06,
  m_CHR = 5e-05,
  strain_transmission = 1,
  strain_seed_date = NULL,
  strain_seed_size = NULL,
  strain_seed_pattern = NULL,
  strain_rel_gamma_E = 1,
  strain_rel_gamma_A = 1,
  strain_rel_gamma_P = 1,
  strain_rel_gamma_C_1 = 1,
  strain_rel_gamma_C_2 = 1,
  strain_rel_p_sympt = 1,
  strain_rel_p_hosp_if_sympt = 1,
  strain_rel_p_icu = 1,
  strain_rel_p_death = 1,
  rel_susceptibility = 1,
  rel_p_sympt = 1,
  rel_p_hosp_if_sympt = 1,
  rel_p_death = 1,
  rel_infectivity = 1,
  vaccine_progression_rate = NULL,
  vaccine_schedule = NULL,
  vaccine_index_dose2 = NULL,
  vaccine_index_booster = NULL,
  vaccine_catchup_fraction = 1,
  n_doses = 2L,
  vacc_skip_progression_rate = 0,
  vacc_skip_from = 1L,
  vacc_skip_to = 1L,
  vacc_skip_weight = 0,
  waning_rate = 0,
  exp_noise = 1e+06,
  cross_immunity = 1,
  n_real_strains = 1L,
  n_vacc_classes = 1L
)
}
\arguments{
\item{start_date}{The start date, as a \code{\link[=sircovid_date]{sircovid_date()}} (i.e.,
the number of days into 2020)}

\item{region}{The region to run the model for. This will bee used
to get population data, which is currently fixed within the
package and is limited to "uk", the four constituent nations
("england", "wales", "scotland", "northern_ireland") and the 7
NHS regions (e.g., "midlands"). These names are case
insensitive.}

\item{beta_date}{A vector of dates (each as a \code{\link[=sircovid_date]{sircovid_date()}})
for changes in beta (the contact rate parameter), or \code{NULL} if
a single value is used for all times (see
\code{\link[=sircovid_parameters_piecewise_linear]{sircovid_parameters_piecewise_linear()}} or
\code{\link[=sircovid_parameters_piecewise_constant]{sircovid_parameters_piecewise_constant()}}, where
this is passed as \code{date}). If \code{beta_type = "piecewise-constant"},
then the first date must be 0.}

\item{beta_value}{A vector of values for beta (the contact rate
parameter). If not given, and if \code{beta_date} is \code{NULL} then a
value of 0.1 will be used through the whole simulation,
otherwise if \code{beta_date} is \code{NULL} this must be a scalar. If
\code{beta_date} is given then \code{beta_date} and \code{beta_value} must have
the same length (see \code{\link[=sircovid_parameters_piecewise_linear]{sircovid_parameters_piecewise_linear()}} or
\code{\link[=sircovid_parameters_piecewise_constant]{sircovid_parameters_piecewise_constant()}}, where this is passed
as \code{value}).}

\item{beta_type}{The type of form used for beta (the contact rate
parameter), which currently can be \code{"piecewise-linear"} or
\code{"piecewise-constant"}}

\item{population}{Population data. A vector of length 17 for the population
size for age groups 0-4, 5-9, ..., 75-79, 80+. If \code{NULL}, the population
data will be sourced within the package for the specified \code{region} (only
if available).}

\item{carehome_beds}{The number of care home beds in the region. If \code{NULL},
this will be sourced within the package for the specified \code{region} (only
if available).}

\item{severity}{Severity data, via Bob Verity's \code{markovid}
package. This needs to be \code{NULL} (use the default bundled data
version in the package), a \link{data.frame} object (for raw severity
data) or a list (for data that has already been processed by
\code{sircovid} for use).  New severity data comes from Bob Verity
via the markovid package, and needs to be carefully calibrated
with the progression parameters.}

\item{progression}{Progression data}

\item{observation}{Either \code{NULL} or a list of observation parameters. If
\code{NULL}, then a list of observation parameters will be generated using
\code{lancelot_parameters_observation(exp_noise)}}

\item{sens_and_spec}{Either \code{NULL} or a list of diagnostic test sensitivity
and specificity parameters. If \code{NULL}, then a list of sensitivity and
specificity parameters will be generated using
\code{lancelot_parameters_sens_and_spec()}}

\item{initial_seed_size}{Initial size of seeding from the S to E
compartment; all seeding is in the 15-19 year old group from
\code{start_date} according to \code{initial_seed_pattern}. The default is 30.}

\item{initial_seed_pattern}{A vector of seeding weights for the initial
seeding. The length represents the number of steps to seed over from the
\code{start_date}, and the \code{initial_seed_size} is split over these steps
according to those weights. If \code{start_date} is not a multiple of the step
size (and thus falls between two steps) then we weight over an additional
step and adjust the weights according to how far the \code{start_date} is from
the previous full step.}

\item{eps}{Change in contact rate for carehome residents}

\item{m_CHW}{Contact rate between carehome workers and either
residents or workers}

\item{m_CHR}{Contact rate between carehome residents}

\item{strain_transmission}{Vector of length two for relative
transmissibility of each strain modelled. Length will define the number of
strains used in the model, either 1 or 2.}

\item{strain_seed_date}{Either \code{NULL} (no seeding) or a
\link{sircovid_date} corresponding to the date the seeding of strain
2 begins.}

\item{strain_seed_size}{Either \code{NULL} (no seeding) or the size of strain 2
seeding from the S to E compartment; all seeding is in the 15-19 year old
group from \code{strain_seed_date} according to \code{strain_seed_pattern}.}

\item{strain_seed_pattern}{Either \code{NULL} (no seeding) or a vector of seeding
weights for the initial seeding. The length represents the number of steps
to seed over from the \code{start_date}, and the \code{strain_seed_size} is split
over these steps according to those weights. If \code{start_seed_date} is not a
multiple of the step size (and thus falls between two steps) then we
weight over an additional step and adjust the weights according to how far
the \code{strain_seed_date} is from the previous full step.}

\item{strain_rel_gamma_E}{Vector of relative rates of progression out of
E (gamma_E) for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{strain_rel_gamma_A}{Vector of relative rates of progression out of
I_A (gamma_A) for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{strain_rel_gamma_P}{Vector of relative rates of progression out of
I_P (gamma_P) for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{strain_rel_gamma_C_1}{Vector of relative rates of progression out of
I_C_1 (gamma_C_1) for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{strain_rel_gamma_C_2}{Vector of relative rates of progression out of
I_C_2 (gamma_C_2) for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{strain_rel_p_sympt}{Vector of relative probabilities of
symptoms for
each strain modelled. If \code{1} all strains have same
probabilities of symptoms. Otherwise vector of same length as
\code{strain_transmission}, where the first value should be 1 (for the first
strain) and subsequent values between 0 and 1. In this case parameters
will be "mirrored" for pseudostrains i.e. the relative probability of
symptoms will be assume the same irrespective of previous infection
with another strain. Alternatively, a vector of twice the length of
\code{strain_transmission} can be provided to allow specifying directly
relative probability of symptoms for each pseudostrain
(with strain 3 - 1.2 and strain 4 = 2.1). To ensure valid
probabilities, p_sympt is upper-truncated at 1 after scaling.}

\item{strain_rel_p_hosp_if_sympt}{Vector of relative probabilities of
hospitalisation given symptoms for
each strain modelled. If \code{1} all strains have same
probabilities of hospitalisation. Otherwise vector of same length as
\code{strain_transmission}, where the first value should be 1 (for the first
strain) and subsequent values between 0 and 1. In this case parameters
will be "mirrored" for pseudostrains i.e. the relative probability of
hospitalisation will be assume the same irrespective of previous infection
with another strain. Alternatively, a vector of twice the length of
\code{strain_transmission} can be provided to allow specifying directly
relative probability of hospitalisation for each pseudostrain
(with strain 3 - 1.2 and strain 4 = 2.1). To ensure valid
probabilities, p_hosp_if_sympt is upper-truncated at 1 after scaling.}

\item{strain_rel_p_icu}{Vector of relative probabilities of
icu given hospitalised for
each strain modelled. If \code{1} all strains have same
probabilities of icu admission. Otherwise vector of same length as
\code{strain_transmission}, where the first value should be 1 (for the first
strain) and subsequent values between 0 and 1. In this case parameters
will be "mirrored" for pseudostrains i.e. the relative probability of
icu admission will be assume the same irrespective of previous infection
with another strain. Alternatively, a vector of twice the length of
\code{strain_transmission} can be provided to allow specifying directly
relative probability of icu admission for each pseudostrain
(with strain 3 - 1.2 and strain 4 = 2.1). To ensure valid
probabilities, p_icu is upper-truncated at 1 after scaling.}

\item{strain_rel_p_death}{Vector of relative probabilities of death for
each strain modelled. If \code{1} all strains have same
probabilities of death. Otherwise vector of same length as
\code{strain_transmission}, where the first value should be 1 (for the first
strain) and subsequent values between 0 and 1. In this case parameters
will be "mirrored" for pseudostrains i.e. the relative probability of
death will be assume the same irrespective of previous infection with
another strain. Alternatively, a vector of twice the length of
\code{strain_transmission} can be provided to allow specifying directly
relative probability of death for each pseudostrain (with strain 3 - 1.2
and strain 4 = 2.1). To ensure valid
probabilities, p_death is upper-truncated at 1 after scaling.}

\item{rel_susceptibility}{A vector or array of values representing the
relative susceptibility of individuals in different vaccination groups.
If a vector, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1. In that case relative
susceptibility will be the same across all age groups within one
vaccination category, and will be the same for all pathogen strains.
Specifying an array instead of a vector allows
different relative susceptibilities by age (first dimension of the array),
pathogen strain (second dimension) and vaccination group
(third dimension); in that case, the first layer (3rd dimension) of
rel_susceptibility should be 1 (for the non-vaccinated group)
for the first column (first infection with first strain) and other
values between 0 and 1}

\item{rel_p_sympt}{A vector or matrix of values of same dimension as
rel_susceptibility representing the
relative probability of symptomatic infection in different
vaccination groups. If a vector, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1.
In that case the relative reduction in probability of symptomatic
infection will be the same across all age groups and all strains
within one vaccination category.
Specifying an array instead of a vector allows different relative
reductions in probability of symptomatic infection by age (first
dimension of the array), pathogen strain (second dimension) and
vaccination group (third dimension); in that case,
the first layer of rel_p_sympt should be 1 (for the non-vaccinated group)
for the first column (first infection with first strain) and other
values between 0 and 1}

\item{rel_p_hosp_if_sympt}{A vector or array of values of same dimension as
rel_susceptibility representing the
relative probability of hospitalisation for symptomatic cases in different
vaccination groups. If a vector, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1.
In that case the relative reduction in probability of hospitalisation for
symptomatic cases will be the same across all age groups and all strains
within one vaccination category.
Specifying an array instead of a vector allows different relative
reductions in probability of hospitalisation for symptomatic cases by age
(first dimension of the array), pathogen strain (second dimension) and
vaccination group (third dimension); in that case,
the first layer of rel_p_hosp_if_sympt should be 1 (for the
non-vaccinated group) for the first column
(first infection with first strain) and other values between 0 and 1}

\item{rel_p_death}{A vector or array of values of same dimension as
rel_susceptibility representing the relative probability of death for
severe cases (either in hospital or the community) in different
vaccination groups. If a vector, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1.
In that case the relative reduction in probability of death for
severe cases will be the same across all age groups and all strains
within one vaccination category.
Specifying an array instead of a vector allows different relative
reductions in probability of death for severe cases by age
(first dimension of the array), pathogen strain (second dimension) and
vaccination group (third dimension); in that case,
the first layer of rel_p_death should be 1 (for the non-vaccinated group)
for the first column (first infection with first strain) and other
values between 0 and 1}

\item{rel_infectivity}{A vector or array of values representing the
relative infectivity of individuals in different vaccination groups,
if they are infected.
If a vector, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1. In that case relative
infectivity will be the same across all age groups and all strains
within one vaccination category.
Specifying an array instead of a vector allows
different relative infectivities by age
(first dimension of the array), pathogen strain (second dimension) and
vaccination group (third dimension); in that case,
the first layer of rel_infectivity should be 1 (for the
non-vaccinated group)
for the first column (first infection with first strain) and other
values between 0 and 1}

\item{vaccine_progression_rate}{A vector or matrix of values of same
dimension as rel_susceptibility representing
the rate of movement between different vaccination classes. If a vector,
it should have as many values as vaccination classes, and the same rates
of progression will be used for all age
groups (the first rate is the vaccination rate and the last of the rates
is the rate of returning to the initial
vacination class); if a matrix, the element on row i and column j is the
rate of progression from the jth vaccination class to the (j+1)th for age
group i.}

\item{vaccine_schedule}{A \link{vaccine_schedule} object indicating the
people to be vaccinated by group over time}

\item{vaccine_index_dose2}{The index to use for the second dose}

\item{vaccine_index_booster}{The index to use for the booster dose}

\item{vaccine_catchup_fraction}{A value between 0 and 1 indicating the
proportion of doses not distributed according to schedule (e.g. because
too many people were in the I or H compartments and could not be
vaccinated at the scheduled time) that we postpone to a later date.
A value of 0 means we do not catch up at all on any missed doses; a
value of 1 means we try to catch up for all missed doses. This is set
to 1 by default}

\item{n_doses}{Number of doses given out, including boosters. Default is
2.}

\item{vacc_skip_progression_rate}{Either a vector of length 19 representing
the base progression rate for "vaccine skip moves" for each group, or a
scalar representing the base progression rate for all groups. Must be 0
(which is the default) if \code{vacc_skip_to == vacc_skip_from}}

\item{vacc_skip_from}{An integer, the vaccine stratum index from which we
allow a "vaccine skip move", which enables individuals to progress through
more than one vaccine class in a single step. Default is 1}

\item{vacc_skip_to}{An integer, the vaccine stratum index representing the
vaccine stratum an individual in stratum \code{vacc_skip_from} can move to in a
"vaccine skip move". Must be either equal to \code{vacc_skip_from} (to allow
the vaccine skip move to be switched "off"), or greater than or equal to
\code{vacc_skip_from + 2} (we already account for moves between successive
vaccine strata). Default is 1}

\item{vacc_skip_weight}{A scalar weight. If movement into \code{vacc_skip_to}
from \code{vacc_skip_to - 1} is controlled by doses, then the "vaccine skip
move" is too, and the weight represents how much those in \code{vacc_skip_from}
are weighted in dose distribution relative to those in \code{vacc_skip_to - 1}.
Must be between 0 and 1. Must be 0 (which is the default) if
\code{vacc_skip_to == vacc_skip_from}}

\item{waning_rate}{A single value or a vector of values representing the
rates of waning of immunity after infection; if a single value the same
rate is used for all age groups; if a vector of values if used it should
have one value per age group.}

\item{exp_noise}{Rate of exponential noise used in the compare
function - typically set to a large value so that noise is small
but non-zero. If set to \code{Inf} then there is no noise in the
observation process (not realistic but useful for testing).}

\item{cross_immunity}{A value or vector of same length as
\code{strain_transmission} that controls the amount of immunity conferred by
previous infection with one strain. If a scalar is given
then same level of cross immunity is assumed between both strains.
Otherwise a vector of length two should be provided where the first value
is the relative protection against infection with strain 2
following infection with strain 1 (i.e. while in the R1 compartment),
and vice versa for the second value. Values between 0 and 1 are allowed
with values of 1 (default) indicating complete cross-immunity,
and values of 0 mean no cross-immunity. Modelling
'superinfections' (being exposed to one strain after
recovering from another) can be turned off by setting
\code{cross_immunity = 1}.}

\item{n_real_strains}{Number of real strains, default is 2.
Value should be '1' (single-strain model) or '2' (multi-strain).}

\item{n_vacc_classes}{Number of vaccine classes. Should match the third
dimension of \code{rel_susceptibility}.}
}
\value{
A list of inputs to the model, many of which are fixed and
represent data. These correspond largely to \code{user()} calls
within the odin code, though some are also used in processing
just before the model is run.
}
\description{
Parameters for the "\link{lancelot}" model.
}
\examples{

region <- "london"
lancelot_parameters(sircovid_date("2020-02-01"), region)

# example set up of vaccination parameters independent of age
# 3 groups: 1) unvaccinated, 2) vaccinated with partial immunity
# 3) fully vaccinated (but with an imperfect vaccine). People return
# to group 1 after a period of time in group 3 to mirror waning immunity

# Assumption: immediately after vaccination susceptibility is reduced by
# 20\%, and then by 50\% when you reach full effect of the vaccination,
# then susceptibility returns to 100\% upon waning of vaccine=induced
# immunity
# effect of vaccination similar across all age groups
rel_susceptibility <- c(1, 0.8, 0.5)

# The vaccine also reduces the risk of symptoms
rel_p_sympt <- c(1, 0.6, 0.3)
# and the risk of hospitalisation for those with symptoms
rel_p_hosp_if_sympt <- c(1, 0.95, 0.95)
# and the risk of death for those with severe disease
rel_p_death <- c(1, 0.9, 0.9)

# The vaccine also reduces infectivity of infected individuals by half
rel_infectivity <- c(1, 0.5, 0.5)

# On average 10000 first doses of vaccine are given every day
# Second doses are given on average 12 weeks after the first dose
# vaccine-induced immunity wanes after a period which is exponentially
# distributed and lasts on average 26 weeks (half a year);
# they are similar across all age groups
vaccine_progression_rate <- c(0, 0, 1/(26*7))

daily_doses <- rep(10000, 365)
mean_days_between_doses <- 12 * 7
n <- vaccine_priority_population(region, uptake = 1)
schedule <- vaccine_schedule_future(
  0, daily_doses, mean_days_between_doses, n)

# generate model parameters
p <- lancelot_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_p_death = rel_p_death,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2,
       n_vacc_classes = 3)

# vaccination parameters are automatically copied across all age groups#
# (and across strains but here we only have 1 strain which is the 2nd
# dimension here)
p$rel_susceptibility
p$rel_p_sympt
p$rel_p_hosp_if_sympt
p$rel_p_death
p$rel_infectivity
# Note that this is only the "base" rate as we fill in the first
# column dynamically based on vaccine_daily_doses
p$vaccine_progression_rate

### same example as above BUT assume a different effect of vaccine in the
### first age group
n_groups <- 19
n_strains <- 1

# Assumption: vaccine is twice more effective at reducing susceptibility
# in the first age group
rel_susceptibility_agegp1 <- c(1, 0.4, 0.25)
rel_susceptibility_other_agegp <- c(1, 0.8, 0.5)
rel_susceptibility <- array(NA, dim = c(n_groups, n_strains, 3))
rel_susceptibility[1, , ] <- rel_susceptibility_agegp1
for (i in seq(2, n_groups)) {
  rel_susceptibility[i, , ] <- rel_susceptibility_other_agegp
}
rel_susceptibility

# But vaccine has the same impact on probability of symptoms and
# hospitalisation for the symptomatic across all age groups
rel_p_sympt <- array(rep(rel_p_sympt, each = n_groups),
  dim = c(n_groups, n_strains, 3))
rel_p_hosp_if_sympt <-
  array(rep(rel_p_hosp_if_sympt, each = n_groups),
  dim = c(n_groups, n_strains, 3))
rel_p_death <-
  array(rep(rel_p_hosp_if_sympt, each = n_groups),
  dim = c(n_groups, n_strains, 3))

# And vaccine has the same impact on onwards infectivity across age groups
rel_infectivity <- array(rep(rel_infectivity, each = n_groups),
  dim = c(n_groups, n_strains, 3))

# the period of build-up of immunity is the same for all age groups,
# lasting on average 2 weeks,
# but the first age group loses immunity more quickly
# (on average after 3 months) than the other age groups
# (on average after 6 months)
vaccine_progression_rate <- cbind(0, 0,
                                  c(1 / (13 * 7),
                                  rep( 1 / (26 * 7), n_groups - 1)))

# generate model parameters
p <- lancelot_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_p_death = rel_p_death,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2,
       n_vacc_classes = 3)

# TODO: add an example of manually set up vaccine schedule

}
