% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_parameters_observation}
\alias{lancelot_parameters_observation}
\title{Lancelot observation parameters}
\usage{
lancelot_parameters_observation(exp_noise = 1e+06)
}
\arguments{
\item{exp_noise}{Rate parameter for the exponentially-distributed noise used
in the likelihood calculation. Typically a large value so noise is small.}
}
\value{
A list of parameter values
}
\description{
Lancelot observation parameters
}
