% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_parameters_progression}
\alias{lancelot_parameters_progression}
\title{Lancelot progression parameters}
\usage{
lancelot_parameters_progression(
  dt,
  gamma_E = NULL,
  gamma_A = NULL,
  gamma_P = NULL,
  gamma_C_1 = NULL,
  gamma_C_2 = NULL,
  gamma_ICU_pre = NULL,
  gamma_ICU_D = NULL,
  gamma_ICU_W_D = NULL,
  gamma_ICU_W_R = NULL,
  gamma_H_D = NULL,
  gamma_H_R = NULL,
  gamma_W_D = NULL,
  gamma_W_R = NULL,
  gamma_G_D = NULL,
  gamma_U = NULL,
  gamma_PCR_pos = NULL
)
}
\arguments{
\item{dt}{The step size}

\item{gamma_E}{Time-varying parameters for the Erlang rate parameter of the
duration in the E (exposed) compartment. See Details.}

\item{gamma_A}{Time-varying parameters for the Erlang rate parameter of the
duration in the I_A (asymptomatic) compartment. See Details.}

\item{gamma_P}{Time-varying parameters for the Erlang rate parameter of the
duration in the I_P (presymptomatic) compartment. See Details.}

\item{gamma_C_1}{Time-varying parameters for the Erlang rate parameter of
the duration in the I_C_1 (first stage symptomatic) compartment. See
Details.}

\item{gamma_C_2}{Time-varying parameters for the Erlang rate parameter of
the duration in the I_C_2 (second stage symptomatic) compartment. See
Details.}

\item{gamma_ICU_pre}{Time-varying parameters for the Erlang rate parameter
of the duration in the ICU_pre (general beds stay before ICU) compartment.
See Details.}

\item{gamma_ICU_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the ICU_D (ICU stay for individuals who die in ICU)
compartment. See Details.}

\item{gamma_ICU_W_D}{Time-varying parameters for the Erlang rate parameter
of the duration in the ICU_W_D (ICU stay for individuals who go on to die
in stepdown) compartment. See Details.}

\item{gamma_ICU_W_R}{Time-varying parameters for the Erlang rate parameter
of the duration in the ICU_W_R (ICU stay for individuals who go on to
recover in stepdown) compartment. See Details.}

\item{gamma_H_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the H_D (general beds stay for individuals who die in
general beds) compartment. See Details.}

\item{gamma_H_R}{Time-varying parameters for the Erlang rate parameter of
the duration in the H_R (general beds stay for individuals who recover in
general beds) compartment. See Details.}

\item{gamma_W_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the W_D (stepdown for individuals who die in stepdown)
compartment. See Details.}

\item{gamma_W_R}{Time-varying parameters for the Erlang rate parameter of
the duration in the W_R (stepdown for individuals who recover in stepdown)
compartment. See Details.}

\item{gamma_G_D}{Time-varying parameters for the Erlang rate parameter of
the duration in the G_D (delay to death in the community/care homes)
compartment. See Details.}

\item{gamma_U}{Time-varying parameters for the Erlang rate parameter of
the duration of the delay from hospital admission to diagnosis for those
not confirmed on admission. Note this duration is a single-stage Erlang.
See Details.}
}
\value{
A list of parameter values
}
\description{
Lancelot progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
\section{Time-varying parameters}{

Every time varying parameter has the same format, which can be \code{NULL} (in
which case a default single value is used) or a list with \code{date} and
\code{value} for changes in the parameter. If \code{value} is scalar then
\code{date} can be \code{NULL} or missing. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.
}

