% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_parameters_sens_and_spec}
\alias{lancelot_parameters_sens_and_spec}
\title{Lancelot sensitivity and specificity parameters}
\usage{
lancelot_parameters_sens_and_spec(
  sero_specificity_1 = 0.9,
  sero_sensitivity_1 = 0.99,
  sero_specificity_2 = 0.9,
  sero_sensitivity_2 = 0.99,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99
)
}
\arguments{
\item{sero_specificity_1}{Specificity of the first serology test assay}

\item{sero_sensitivity_1}{Sensitivity of the first serology test assay}

\item{sero_specificity_2}{Specificity of the second serology test assay}

\item{sero_sensitivity_2}{Sensitivity of the second serology test assay}

\item{pillar2_specificity}{Specificity of the Pillar 2 test}

\item{pillar2_sensitivity}{Sensitivity of the Pillar 2 test}

\item{react_specificity}{Specificity of the REACT test}

\item{react_sensitivity}{Sensitivity of the REACT test}
}
\value{
A list of parameter values
}
\description{
Lancelot observation parameters
}
