% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot.R
\name{lancelot_prepare_data}
\alias{lancelot_prepare_data}
\title{Prepare data for particle filter}
\usage{
lancelot_prepare_data(data)
}
\arguments{
\item{data}{A data.frame of data}
}
\value{
A data frame
}
\description{
Prepare data for the particle filter
}
