% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancelot_rt.R
\name{lancelot_Rt_trajectories}
\alias{lancelot_Rt_trajectories}
\title{Compute Rt for a set of trajectories}
\usage{
lancelot_Rt_trajectories(
  step,
  S,
  pars,
  prob_strain = NULL,
  initial_step_from_parameters = TRUE,
  shared_parameters = NULL,
  type = NULL,
  interpolate_every = NULL,
  interpolate_critical_dates = NULL,
  interpolate_min = NULL,
  eigen_method = "power_iteration",
  R = NULL,
  weight_Rt = FALSE,
  keep_strains_Rt = FALSE
)
}
\arguments{
\item{step}{A vector of steps}

\item{S}{A 3d ((n groups x n vaccine classes) x n trajectories x n steps)
array of "S" compartment counts}

\item{pars}{Either a single \code{\link[=lancelot_parameters]{lancelot_parameters()}} object
(shared parameters) or an unnamed list of
\code{\link[=lancelot_parameters]{lancelot_parameters()}} objects, the same length as \code{ncol(S)}.}

\item{prob_strain}{A 3d ((n groups x n strains) x n trajectories x n steps)
array of "prob_strain" model outputs. Default is \code{NULL}, but it must be
specified if there is more than one strain.}

\item{initial_step_from_parameters}{If \code{TRUE}, then \code{step[[1]]} is
replaced by the value of \code{initial_step} from the parameters.
This is usually what you want. (From sircovid 0.12.13 this
parameter means "initial step is zero" and will probably be
updated in a future version).}

\item{shared_parameters}{Should \code{pars} be treated as a single
shared list? Leave as \code{NULL} to detect automatically, set to
\code{TRUE} or \code{FALSE} to force it to be interpreted one way or the
other which may give more easily interpretable error messages.}

\item{type}{A character vector of possible Rt types to
compute. Can be any or all of \code{eff_Rt_all}, \code{eff_Rt_general},
\code{Rt_all} and \code{Rt_general}}

\item{interpolate_every}{Spacing (in days) to use between interpolated
points}

\item{interpolate_critical_dates}{Optional vector of critical sircovid
dates to use when interpolating. Interpolation will be done in
blocks between the first time step of these dates, with each block
starting on the first time step of the date given.
So if you give a \code{interpolate_critical_dates} of \code{c(20, 50)} then blocks \emph{start} on first time step of days 20 and 50,
i.e.: \verb{[1, 20)}, \verb{[20, 50)}, \verb{[50, end]}.}

\item{interpolate_min}{The minimum number of steps to include
within a block. If there are fewer points than this then all
points are used (i.e., no interpolation is done) or
\code{interpolate_every} is reduced until at least this many points
were used. This can be used to specify a lower bound on the
error of small regions. If Rt is small it won't matter that
much. You do need to specify something though or interpolation
will not happen, and do not use less than 3 as we use spline
interpolation and that will not work with fewer than 3 points.}

\item{eigen_method}{The eigenvalue method to use (passed to
\link[eigen1:eigen1]{eigen1::eigen1} as \code{method})}

\item{R}{A 3d ((n groups x n strains x n vaccine classes) x
n trajectories x n steps) array of "R" compartment counts, required for
multi-strain models.}

\item{weight_Rt}{If \code{TRUE} then computes the weighted average
of the Rt for all strains, otherwise all calculations are returned with
an additional dimension to index each strain.}

\item{keep_strains_Rt}{Additional argument for when \code{weight_Rt} is \code{TRUE}
(has no impact otherwise). If \code{TRUE}, then the Rt for each strain is
returned along with the weighted average, otherwise just the weighted
average is returned. When \code{TRUE}, the dimension indexing these lists
strains first, and then the weighted average.}
}
\value{
As for \code{\link[=lancelot_Rt]{lancelot_Rt()}}, except that every element is a
matrix, not a vector.
}
\description{
Compute "Rt" for a set of simulated trajectories (e.g., the result
of the \verb{$iterate()} method of \link{lancelot}, \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}} or
\code{\link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict()}}. The trajectories may or may not share
parameters.
}
