% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{multiply_odds}
\alias{multiply_odds}
\title{Multiply probability on odds scale}
\usage{
multiply_odds(prob, multiplier)
}
\arguments{
\item{prob}{Probability to multiply}

\item{multiplier}{Multiplier}
}
\description{
Multiply a probability on the odds scale with a given multiplier.
}
