% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Create parameters}
\usage{
parameters(
  transmission_model = "POLYMOD",
  country = "United Kingdom",
  age_limits = c(0, 10, 20, 30, 40, 50, 60, 70, 80),
  progression_parameters = "SPI-M-Feb-2009",
  beta = 0.1,
  survey_pop = NULL,
  output_parameter_table = TRUE
)
}
\arguments{
\item{transmission_model}{Model type}

\item{country}{Country name}

\item{age_limits}{Vector of age}

\item{progression_parameters}{Progression parameters}

\item{beta}{Beta, obvs.}

\item{survey_pop}{A survey population perhaps}

\item{output_parameter_table}{The table of output parameters}
}
\description{
Create parameters for use with the model
}
