% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{particle_filter}
\alias{particle_filter}
\title{Run a particle filter}
\usage{
particle_filter(
  data,
  pars_model,
  pars_obs,
  n_particles,
  start_date,
  time_steps_per_day,
  forecast_days = 0,
  output_states = FALSE,
  save_particles = FALSE
)
}
\arguments{
\item{data}{data}

\item{pars_model}{parameters of the model}

\item{pars_obs}{parameters of the observations}

\item{n_particles}{number of particles}

\item{start_date}{start date}

\item{time_steps_per_day}{Number of timesteps per day}

\item{forecast_days}{number of days to forecast forward from end states}

\item{output_states}{Logical, indicating if output of states is wanted}

\item{save_particles}{Logical, indicating if we save full particle histories}
}
\description{
Run a particle filter
}
