% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{regions}
\alias{regions}
\title{Regions}
\usage{
regions(type)
}
\arguments{
\item{type}{The name of a region type; must be one of "all",
"england", or "nations"}
}
\value{
A character vector
}
\description{
Generate lists of regions that we use for various tasks.
}
\examples{
sircovid::regions("england")
}
