% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{rel_severity_params}
\alias{rel_severity_params}
\title{Print relative severity parameters}
\usage{
rel_severity_params()
}
\value{
A character vector with the names of the relative severity
parameters
}
\description{
Return relative severity parameters in a specified order
}
