% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{reorder_sample}
\alias{reorder_sample}
\title{Reorder samples}
\usage{
reorder_sample(sample, rank)
}
\arguments{
\item{sample}{An \code{mcstate_pmcmc} object}

\item{rank}{A vector of ranks to reorder by}
}
\value{
An \code{mcstate_pmcmc} object with appropriately reordered elements
}
\description{
Reorder samples according to a predefined ranking
}
