% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rotate_strains}
\alias{rotate_strains}
\title{Rotate strains}
\usage{
rotate_strains(state, info, ...)
}
\arguments{
\item{state}{Model state}

\item{info}{Model info}

\item{...}{Additional arguments, ignored. This exists so that this
function can be used as an argument to
\link[mcstate:multistage_epoch]{mcstate::multistage_epoch}.  Practically your two model
informations in this case would be equivalent.}
}
\description{
Rotate strains, so that strain 1 becomes the sum of strains 1 and
2 and strain 2 is empty. Use this to allow sequential replacement
of strains.
}
