% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircovid.R
\name{run_particle_filter}
\alias{run_particle_filter}
\title{Run particle filter}
\usage{
run_particle_filter(case_data_file, admissions_data_file,
  itu_colname = "number_of_confirmed_covid_19_patients_in_hdu_itu_at_0800_total",
  itu_ignore_dates = c("2020-03-15"), first_data_date = "2020-03-01",
  model_start_date = "2020-02-02", country_name = "United_Kingdom",
  case_file_day_shift = 2, model_params, obs_params = list(phi_ICU =
  0.95, k_ICU = 2, phi_death = 1789/1651, k_death = 2, exp_noise = 1e+06),
  n_particles = 1000)
}
\arguments{
\item{case_data_file}{CSV file containing COVID-19 cases, from
the ECDC website}

\item{admissions_data_file}{Admission totals, from ncov-outputs}

\item{itu_colname}{Name of the column in the time series file
to be used for fitting ITU cases}

\item{itu_ignore_dates}{List of dates to discard from the
data due to error or otherwise}

\item{first_data_date}{Date to start data fit from}

\item{model_start_date}{Date to run model simulations from}

\item{country_name}{Country name in case reports}

\item{case_file_day_shift}{Number of days to subtract from
dateRep in case file by to match with date in 
\code{time_series_file}}

\item{model_params}{Parameters to run model simulations,
generated by \code{generate_parameters()}}

\item{obs_params}{List of parameters used for comparing
model to data in the particle filter}

\item{n_particles}{Number of particles}
}
\description{
Create a model, and fit with the particle filter
}
