% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\name{sample_pmcmc}
\alias{sample_pmcmc}
\title{Sample pmcmc}
\usage{
sample_pmcmc(
  mcmc_results,
  burn_in = 1,
  n_sample = 10,
  n_particles = 100,
  forecast_days = 0
)
}
\arguments{
\item{mcmc_results}{Output of \code{\link{pmcmc}}.}

\item{burn_in}{Number of burn-in samples to discard}

\item{n_sample}{Number of parameter pairs to be sampled. This will 
determine how many trajectories are returned. Integer. Default = 10. This 
will determine how many trajectories are returned.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}

\item{forecast_days}{Number of days being forecast. Default = 0}
}
\value{
\code{\link{list}}. First element (trajectories) is a 3 
  dimensional array of trajectories (time, state, tranjectories). Second 
  element (param_grid) is the parameters chosen when  sampling from the 
  \code{scan_results} grid and the third dimension (inputs) is a list of
  model inputs.
}
\description{
Take a parameter search produced by \code{\link{pmcmc}} and 
sample \code{n_sample} from the parameter space
on their probability. For each parameter pair chosen, run particle
filter with \code{num_particles} and sample 1 trajectory
}
