% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{scan_beta_date}
\alias{scan_beta_date}
\title{Grid search of beta and start date}
\usage{
scan_beta_date(
  min_beta,
  max_beta,
  beta_step,
  first_start_date,
  last_start_date,
  day_step,
  data,
  model_fun = NULL,
  model_params = NULL,
  compare_fun = NULL,
  pars_obs = NULL,
  n_particles = 100
)
}
\arguments{
\item{min_beta}{Minimum value of beta in the search}

\item{max_beta}{Maximum value of beta in the search}

\item{beta_step}{Step to increment beta between min and max}

\item{first_start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{last_start_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{data}{Hosptial data to fit to. See \code{example.csv}
and \code{particle_filter_data()}}

\item{model_fun}{An odin function used to create the model.
If NULL, uses basic.}

\item{model_params}{Model parameters, from a call to
\code{generate_parameters()}. If NULL, uses defaults as
in unit tests.}

\item{compare_fun}{function used to calculate the log weights
in the particle filter. If NULL, uses compare_icu.}

\item{pars_obs}{list of parameters to use for the compare_fun.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over beta and start date.
This is parallelised, first run \code{plan(multiprocess)} to set 
this up.
}
