% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{scan_beta_date_reduction}
\alias{scan_beta_date_reduction}
\title{Grid search of beta and start date}
\usage{
scan_beta_date_reduction(
  min_beta,
  max_beta,
  beta_step,
  first_start_date,
  last_start_date,
  day_step,
  min_beta_reduction,
  max_beta_reduction,
  beta_reduction_step,
  data,
  sircovid_model = basic_model(),
  model_params = NULL,
  pars_obs = NULL,
  n_particles = 100
)
}
\arguments{
\item{min_beta}{Minimum value of beta in the search}

\item{max_beta}{Maximum value of beta in the search}

\item{beta_step}{Step to increment beta between min and max}

\item{first_start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{last_start_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{min_beta_reduction}{Minimum value of beta_reduction in the search}

\item{max_beta_reduction}{Maximum value of beta_reduction in the search}

\item{beta_reduction_step}{Step to increment beta_reduction between min and max}

\item{data}{Hosptial data to fit to. See \code{example.csv}
and \code{particle_filter_data()}}

\item{sircovid_model}{An odin model generator and comparison function.}

\item{model_params}{Model parameters, from a call to
\code{generate_parameters()}. If NULL, uses defaults as
in unit tests.}

\item{pars_obs}{list of parameters to use for the comparison function.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over beta, start date and
beta_reduction. This is parallelised, first run \code{plan(multiprocess)}
to set this up.
}
