% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{scan_beta_lockdown_date}
\alias{scan_beta_lockdown_date}
\title{Grid search of beta and start date}
\usage{
scan_beta_lockdown_date(
  min_beta_start,
  max_beta_start,
  beta_start_step,
  min_beta_end,
  max_beta_end,
  beta_end_step,
  first_start_date,
  last_start_date,
  day_step,
  data,
  sircovid_model = basic_model(),
  model_params = NULL,
  pars_obs = NULL,
  n_particles = 100
)
}
\arguments{
\item{min_beta_start}{Minimum value of beta_start in the search}

\item{max_beta_start}{Maximum value of beta_start in the search}

\item{beta_start_step}{Step to increment beta_start between min and max}

\item{min_beta_end}{Minimum value of beta_end in the search}

\item{max_beta_end}{Maximum value of beta_end in the search}

\item{beta_end_step}{Step to increment beta_end between min and max}

\item{first_start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{last_start_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{data}{Hosptial data to fit to. See \code{example.csv}
and \code{particle_filter_data()}}

\item{sircovid_model}{An odin model generator and comparison function.}

\item{model_params}{Model parameters, from a call to
\code{generate_parameters()}. If NULL, uses defaults as
in unit tests.}

\item{pars_obs}{list of parameters to use for the comparison function.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over beta, start date and
beta_reduction. This is parallelised, first run \code{plan(multiprocess)}
to set this up.
}
