% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircovid.R
\name{sircovid}
\alias{sircovid}
\title{Create a model}
\usage{
sircovid(params)
}
\arguments{
\item{params}{List of parameters of the model. Get
with \code{generate_parameters}}
}
\description{
Create a model
}
