% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sircovid_data}
\alias{sircovid_data}
\title{Prepare data for mcstate}
\usage{
sircovid_data(data, start_date, dt, expected = NULL)
}
\arguments{
\item{data}{A \code{data.frame} object suitable for
\code{\link[mcstate:particle_filter_data]{mcstate::particle_filter_data()}}}

\item{start_date}{The start date, as a \code{\link[=sircovid_date]{sircovid_date()}}, R "Date"
object or a string in ISO 8601 format (YYYY-MM-DD)}

\item{dt}{The time step (fraction of a day that each step
represents) as used to create the model object}

\item{expected}{Optional list of expected column names and types
that are expected (use \code{NA_real_}). These will be filled in with
the value where missing.}
}
\value{
A data.frame suitable for use with \code{mcstate} functions
such as \code{\link[mcstate:particle_filter]{mcstate::particle_filter()}} and \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}}
}
\description{
Convert data into mcstate format; this is a thin wrapper around
\code{\link[mcstate:particle_filter_data]{mcstate::particle_filter_data()}} which adds a dummy step in front
of the first data point so that we can use the previous state and
the current states to convert cumulative measures into net daily
changes.
}
\examples{
# A data sert that has data from the first of February to the first of
# March (one column of data called 'x')
from <- as.Date("2020-02-01")
to <- as.Date("2020-03-01")
d <- data.frame(date = seq(from, to, by = 1),
                x = runif(to - from + 1),
                stringsAsFactors = FALSE)

# Get this ready for sircovid/mcstate assuming the seeding starts on
# the 15th of January and we take 4 steps per day.
sircovid_data(d, start_date = "2020-01-15", 1 / 4)
}
