% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{sircovid_date}
\alias{sircovid_date}
\title{Convert date to number}
\usage{
sircovid_date(date)
}
\arguments{
\item{date}{A Date object, or something that can be converted to one}
}
\value{
An integer, being the number of days into 2020
}
\description{
Convert a date into the number of days into (or past) 2020
}
\examples{
sircovid_date("2020-01-01")
sircovid_date(c("2020-03-01", "2020-10-01"))
}
