% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sircovid_date_as_Date}
\alias{sircovid_date_as_Date}
\title{sircovid_date_as_Date}
\usage{
sircovid_date_as_Date(sircovid_date)
}
\arguments{
\item{sircovid_date}{numeric days since the start of 2020 (i.e. 2020-01-01 = 1)}
}
\value{
\code{date}
}
\description{
Convert a date from the representation used in the sircovid package,
which is days since the start of 2020 (i.e. 2020-01-01 = 1) into 
a lubridate
}
