% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sircovid_parameters_severity}
\alias{sircovid_parameters_severity}
\title{Process severity data}
\usage{
sircovid_parameters_severity(params)
}
\arguments{
\item{params}{Severity data, via Bob Verity's \code{markovid}
package. This needs to be \code{NULL} (use the default bundled data
version in the package), a \link{data.frame} object (for raw severity
data) or a list (for data that has already been processed by
\code{sircovid} for use).  New severity data comes from Bob Verity
via the markovid package, and needs to be carefully calibrated
with the progression parameters.}
}
\value{
A list of age-structured probabilities
}
\description{
Process severity data
}
