% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sircovid_parameters_step_expand}
\alias{sircovid_parameters_step_expand}
\title{Expand beta steps}
\usage{
sircovid_parameters_step_expand(step, beta_step)
}
\arguments{
\item{step}{A vector of steps}

\item{beta_step}{A vector of betas}
}
\value{
A numeric vector the same length as \code{step}
}
\description{
Expand \code{beta_step} based on a series of \code{step}s.  Use this to
convert between the values passed to
\code{\link[=sircovid_parameters_piecewise_linear]{sircovid_parameters_piecewise_linear()}} and the actual beta
values for a given set of steps.
}
