% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sircovid_simulate}
\alias{sircovid_simulate}
\title{Simulate a sircovid model}
\usage{
sircovid_simulate(
  mod,
  state,
  p,
  events,
  index = NULL,
  n_threads = 1L,
  seed = NULL
)
}
\arguments{
\item{mod}{The model object to simulate from}

\item{state}{A matrix of state; rows correspond to state
variables, columns correspond to different parameter sets or
realisations.}

\item{p}{An unnamed list of parameter objects; must be the same
length as \code{state} has columns.}

\item{events}{A \code{\link[=sircovid_simulate_events]{sircovid_simulate_events()}} object with "events"
data, representing changes in the parameters, and the overall
timescale of the simulation.}

\item{index}{A (possibly named) integer vector representing an
index into the model state that should be returned; will be set
into the model with \verb{$set_index}}

\item{n_threads}{The number of threads to use in the
simulation. This only has an effect if your sircovid was built
with OpenMP support (possibly not the case on macOS).  Each
simulation will be run on potentially a different thread.}

\item{seed}{Optional seed to create the model with}
}
\value{
A 3-d array with dimensions representing state (along
\code{index}), parameter set (along \code{p}) and time. The array has
attributes \code{step} and \code{date} corresponding to the step (relative
to 2020-01-01) and date.
}
\description{
Simulate from a sircovid model (\link{basic} or \link{lancelot}) given a
set of starting positions and parameters. This forms the core of
our projections support, and is typically used after fitting a
model.
}
