% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sircovid_simulate_events}
\alias{sircovid_simulate_events}
\title{Simulation events}
\usage{
sircovid_simulate_events(date_start, date_end, data)
}
\arguments{
\item{date_start}{The start date for the simulation, as an ISO
date or \link{sircovid_date}}

\item{date_end}{The end date for the simulation, as an ISO
date or \link{sircovid_date}}

\item{data}{Events data; this must be a named list, with each name
being in ISO 8601 (YYYY-MM-DD) format, and value being a named
set of parameters to apply to our base set. So a value of
\code{list("2021-03-01" = list(vaccine_daily_doses = 10000))} would
represent a change to the \code{vaccine_daily_doses} parameter from
the 1st of March 2021.  Each entry can change as many parameters
as wanted. These changes are \emph{not cumulative}, so if you want a
change to persist across multiple events it must exist in each
of these events.}
}
\value{
A \code{sircovid_simulate_events} object; consider its contents
opaque for now as these may change.
}
\description{
Events for use with \code{\link[=sircovid_simulate]{sircovid_simulate()}}. This captures the
overall timescape of the simulation and a set of "events" which
correspond to changes in parameters.
}
