% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_squared.R
\name{smc_squared}
\alias{smc_squared}
\title{Run an SMC^2 algorithm}
\usage{
smc_squared(
  data,
  sircovid_model,
  model_params,
  beta_params,
  pars_obs,
  pars_seeding,
  earliest_seeding_date,
  n_particles,
  n_param_sets,
  fitted_params,
  degeneracy_threshold,
  covariance_scaling
)
}
\arguments{
\item{data}{Data to fit to.}

\item{sircovid_model}{An odin model generator and comparison function}

\item{model_params}{Default set of parameters used in odin_model}

\item{beta_params}{List of parameters used for time-varying beta}

\item{pars_obs}{Observation parameters}

\item{pars_seeding}{Seeding parameters}

\item{earliest_seeding_date}{Earliest possible date for seeding}

\item{n_particles}{Number of particles}

\item{n_param_sets}{Number of parameter sets}

\item{fitted_params}{list of fitted parameters, with a list of 
log-density function for prior ($dprior), random generator for 
prior ($rprior) and random generator used for proposal ($rproposal)}

\item{degeneracy_threshold}{A number in (0,1), threshold used to 
determine whether or not to do an MCMC move for the parameters}

\item{covariance_scaling}{Scaling of the weighted covariance for 
multivariate normal proposal of parameters}
}
\description{
Run an SMC^2 algorithm
}
