% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasts.R
\name{sum_over_compartments}
\alias{sum_over_compartments}
\title{Summarise trajectories}
\usage{
sum_over_compartments(sample_grid_res)
}
\arguments{
\item{object}{Results from \code{\link{sample_grid_scan}}}
}
\description{
Take a sampled grid search produced by \code{\link{sample_grid_scan}} summarise trajectories 
over compartmets
}
