% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid.R
\name{summary.sample_grid_search}
\alias{summary.sample_grid_search}
\title{Summarise grid forecast}
\usage{
\method{summary}{sample_grid_search}(
  object,
  ...,
  what = "deaths",
  quantiles = seq(from = 0.05, to = 0.95, by = 0.05)
)
}
\arguments{
\item{object}{Results from \code{\link{sample_grid_scan}}}

\item{...}{other arguments to \code{summary()}}

\item{what}{Output to summaries. "deaths", "icu" or "hosp"}

\item{quantiles}{Quantiles to summarise forecast variance.}
}
\description{
Take a sampled grid search produced by \code{\link{sample_grid_scan}} and 
extract forecasts and quantiles in text format
}
