% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{upgrade_state}
\alias{upgrade_state}
\title{Upgrade model state after change}
\usage{
upgrade_state(state_orig, info_orig, info_new, allowed = NULL)
}
\arguments{
\item{state_orig}{Old model info}

\item{info_orig}{Old model info}

\item{info_new}{New model info}

\item{allowed}{Character vector of allowed states to add; new
values will be initialised as zeros.}
}
\value{
A new copy of the state
}
\description{
Upgrade model state
}
