% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccination_remap_state}
\alias{vaccination_remap_state}
\title{Remap state to include vaccination}
\usage{
vaccination_remap_state(state_orig, info_orig, info_vacc)
}
\arguments{
\item{state_orig}{An original state matrix (rows representing
state, columns representing particles or samples). This must be
the complete model state.}

\item{info_orig}{The results of \code{info()} from the model object
without vaccination}

\item{info_vacc}{The results of \code{info()} from the model object
with vaccination}
}
\value{
A 3d array with more rows than \code{state}.
}
\description{
Remap state that was run with one stratum to support
vaccination. This is useful where fitting was done pre-vaccination
and the model state needs to be expanded to support vaccination
for simulations.
}
