% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccination_schedule_future}
\alias{vaccination_schedule_future}
\title{Create vaccination schedule}
\usage{
vaccination_schedule_future(
  daily_doses_value,
  daily_doses_date,
  mean_days_between_doses,
  priority_population
)
}
\arguments{
\item{daily_doses_value}{A vector of doses per day}

\item{daily_doses_date}{A \link{sircovid_date} object corresponding to
the first date in daily_doses_value}

\item{mean_days_between_doses}{Assumed mean days between doses one
and two}

\item{priority_population}{Output from
\link{vaccination_priority_population}, giving the number of people
to vaccinate in each age (row) and priority group (column)}
}
\description{
Create future vaccination schedule from a projected number of
daily doses.
}
