% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_eligibility}
\alias{vaccine_eligibility}
\title{Create vaccine eligibility vector}
\usage{
vaccine_eligibility(min_age)
}
\arguments{
\item{min_age}{Minimum age to be eligible for vaccination
(includes this age).}
}
\value{
A vector of length 19
}
\description{
Create a vector of vaccine eligibility, based on a minimum
age. The assumption is that everyone below that age is ineligible
and everyone including and above is eligibility.  This will
practically be modified further by assumptions of uptake.  Note
that carehome residents and workers are \strong{always} eligible for
vaccination, regardless of the minimum age.
}
