% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_priority_proportion}
\alias{vaccine_priority_proportion}
\alias{vaccine_priority_population}
\title{Compute vaccination order}
\usage{
vaccine_priority_proportion(
  uptake,
  prop_hcw = NULL,
  prop_very_vulnerable = NULL,
  prop_underlying_condition = NULL
)

vaccine_priority_population(
  region,
  uptake,
  prop_hcw = NULL,
  prop_very_vulnerable = NULL,
  prop_underlying_condition = NULL,
  carehomes = TRUE
)
}
\arguments{
\item{uptake}{A vector of length 19 with fractional uptake per
group. If a single number is given it is shared across all
groups (note that this includes under-18s)}

\item{prop_hcw}{Assumed fraction of healthcare workers in each
group (length 19) - if \code{NULL} we use a default that is a guess
with hopefully the right general shape.}

\item{prop_very_vulnerable}{Assumed fraction "very vulnerable" in
each group (length 19) - if \code{NULL} we use a default that is a
guess with hopefully the right general shape.}

\item{prop_underlying_condition}{Assumed fraction "underlying
condition" in each group (length 19) - if \code{NULL} we use a
default that is a guess with hopefully the right general shape.}

\item{region}{Region to use to get total population numbers}
}
\value{
A matrix with n_groups rows (19) and columns representing
priority groups, and element (i, j) is the proportion in group i
who should be vaccinated as part of priority group j, accounting
for uptake so that the sum over the rows corresponds to the
total fractional uptake in that group. For
\code{vaccine_priority_population}, the total number of
individuals replaces the proportion (based on the demography
used by sircovid).
}
\description{
Compute vaccination priority following JCVI ordering.
}
\details{
https://tinyurl.com/8uwtatvm

Assmuing independance between job (e.g. HCW) and clinical condition

But assuming one is either counted as "clinically extremely
vulnerable" or "with underlying health conditions" but not both

The JCVI priority groups, in decending order are:
\enumerate{
\item residents in a care home for older adults and their carers
\item all those 80 years of age and over and frontline health and social
care workers
\item all those 75 years of age and over
\item all those 70 years of age and over and clinically extremely
vulnerable individuals
\item all those 65 years of age and over
\item all individuals aged 16 years to 64 years with underlying health
conditions which put them at higher risk of serious disease and
mortality
\item all those 60 years of age and over
\item all those 55 years of age and over
\item all those 50 years of age and over
\item all those 40-49 years of age and over
\item all those 30-39 years of age and over
\item all those 18-29 years of age and over
}
}
