% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule}
\alias{vaccine_schedule}
\title{Create vaccine schedule}
\usage{
vaccine_schedule(date, doses, n_doses = 2L)
}
\arguments{
\item{date}{A single date, representing the first day that
vaccines will be given}

\item{doses}{A 3d array of doses representing (1) the model group
(19 rows for the carehomes model), (2) the dose (must be length
2 at present) and (3) time (can be anything nonzero). The values
represent the number of vaccine doses in that group for that
dose for that day. So for \code{doses[i, j, k]} then it is for the
ith group, the number of jth doses on day \code{(k - 1) + date}}

\item{n_doses}{The number of doses in the schedule. Typically (and
by default) this will be 2, but if using booster doses 3 (and in
future we may extend further).}
}
\value{
A \code{vaccine_schedule} object
}
\description{
Create a vaccine schedule for use with \link{carehomes_parameters}
}
