% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_from_data}
\alias{vaccine_schedule_from_data}
\title{Create historical vaccine schedule}
\usage{
vaccine_schedule_from_data(data, region, uptake)
}
\arguments{
\item{data}{A data.frame with columns \code{date}, \code{age_band_min},
\code{dose1} and \code{dose2}.}
}
\value{
A \link{vaccine_schedule} object
}
\description{
Create a historical vaccine schedule from data
}
