% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_future}
\alias{vaccine_schedule_future}
\title{Create vaccination schedule}
\usage{
vaccine_schedule_future(
  start,
  daily_doses_value,
  mean_days_between_doses,
  priority_population,
  lag_groups = NULL,
  lag_days = NULL,
  booster_daily_doses_value = NULL,
  booster_proportion = rep(1L, 19)
)
}
\arguments{
\item{start}{Either a \link{sircovid_date} object corresponding to the
first date in daily_doses_value, or a \link{vaccine_schedule} object
corresponding to previously carried out vaccination.}

\item{daily_doses_value}{A vector of doses per day.}

\item{mean_days_between_doses}{Assumed mean days between doses one
and two}

\item{priority_population}{Output from
\link{vaccine_priority_population}, giving the number of people
to vaccinate in each age (row) and priority group (column)}

\item{lag_groups}{Row indices, corresponding to age
groups in which a lag should be added to the start time of the dose
schedule returned by \link{vaccine_schedule}, if NULL then no lag is added.
Ignored if \code{lag_groups} is NULL.}

\item{lag_days}{If \code{lag_groups} is not NULL then specifies the number of
days to add the start of the dose schedule for the given groups. Ignored
if \code{lag_groups} is NULL.}

\item{booster_daily_doses_value}{A vector of booster doses per day.}

\item{booster_proportion}{Proportion of the groups in
\code{priority_population} to boost, default is all groups; ignored if
\code{booster_daily_doses_value} is NULL.}
}
\description{
Create future vaccination schedule from a projected number of
daily doses.
}
