% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_scenario}
\alias{vaccine_schedule_scenario}
\title{High-level vaccine scenario creation}
\usage{
vaccine_schedule_scenario(
  schedule_past,
  doses_future,
  end_date,
  mean_days_between_doses,
  priority_population
)
}
\arguments{
\item{schedule_past}{A \link{vaccine_schedule} object corresponding to
previously carried out vaccination.}

\item{doses_future}{A named vector of vaccine doses to give in the
future. Names must be in ISO date format.}

\item{end_date}{The final day in the future to create a schedule
for. After this date the model will assume 0 vaccine doses given
so an overestimate is probably better than an underestimate.}

\item{mean_days_between_doses}{Assumed mean days between doses one
and two}

\item{priority_population}{Output from
\link{vaccine_priority_population}, giving the number of people
to vaccinate in each age (row) and priority group (column)}
}
\value{
A \link{vaccine_schedule} object
}
\description{
Create a vaccination scenario
}
