% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{zero_boundary}
\alias{zero_boundary}
\title{Check if the boundaries of a matrix are within the specified
tolerance of 0.}
\usage{
zero_boundary(array, tolerance)
}
\arguments{
\item{array}{Probability matrix whose edges need to be checked.
Can be a matrix or an array.}

\item{tolerance}{Edges of matrix should be at least this much away
from 0.}
}
\value{
TRUE if the each edge of array is within specified toleance
of 0, FALSE otherwise
}
\description{
Check if the boundaries of a matrix are within the specified
tolerance of 0.
}
\author{
Sangeeta Bhatia
}
