% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_compare}
\alias{basic_compare}
\title{Compare observed and modelled data for the basic model}
\usage{
basic_compare(state, prev_state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current day. This is
assumed to be filtered following \code{\link[=basic_index]{basic_index()}} so contains
rows corresponding to itu and deaths.}

\item{prev_state}{State vector for the end of the previous day, as
for \code{state}.}

\item{observed}{Observed data. This will be a list with elements
\code{itu} (number of itu/icu beds occupied) and \code{deaths} (number of
deaths over this day).}

\item{pars}{A list of observation parameters, as created by
\code{\link[=basic_parameters_observation]{basic_parameters_observation()}}}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the basic model. This
conforms to the mcstate interface.
}
\examples{
state <- rbind(10:15, 1:6) # ICU, D
prev_state <- matrix(1, 2, 6)
observed <- list(itu = 13, deaths = 3)
pars <- basic_parameters_observation()
basic_compare(state, prev_state, observed, pars)
basic_compare(state * 5, prev_state, observed, pars)
}
