% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic_parameters}
\alias{basic_parameters}
\title{Parameters for the basic model}
\usage{
basic_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  beta_type = "piecewise-linear",
  severity = NULL,
  exp_noise = 1e+06,
  initial_seed_size = 30,
  initial_seed_pattern = 1
)
}
\arguments{
\item{start_date}{The start date, as a \code{\link[=sircovid_date]{sircovid_date()}} (i.e.,
the number of days into 2020)}

\item{region}{The region to run the model for. This will bee used
to get population data, which is currently fixed within the
package and is limited to "uk", the four constituent nations
("england", "wales", "scotland", "northern_ireland") and the 7
NHS regions (e.g., "midlands"). These names are case
insensitive.}

\item{beta_date}{A vector of dates (each as a \code{\link[=sircovid_date]{sircovid_date()}})
for changes in beta (the contact rate parameter), or \code{NULL} if
a single value is used for all times (see
\code{\link[=sircovid_parameters_piecewise_linear]{sircovid_parameters_piecewise_linear()}} or
\code{\link[=sircovid_parameters_piecewise_constant]{sircovid_parameters_piecewise_constant()}}, where
this is passed as \code{date}). If \code{beta_type = "piecewise-constant"},
then the first date must be 0.}

\item{beta_value}{A vector of values for beta (the contact rate
parameter). If not given, and if \code{beta_date} is \code{NULL} then a
value of 0.1 will be used through the whole simulation,
otherwise if \code{beta_date} is \code{NULL} this must be a scalar. If
\code{beta_date} is given then \code{beta_date} and \code{beta_value} must have
the same length (see \code{\link[=sircovid_parameters_piecewise_linear]{sircovid_parameters_piecewise_linear()}} or
\code{\link[=sircovid_parameters_piecewise_constant]{sircovid_parameters_piecewise_constant()}}, where this is passed
as \code{value}).}

\item{beta_type}{The type of form used for beta (the contact rate
parameter), which currently can be \code{"piecewise-linear"} or
\code{"piecewise-constant"}}

\item{severity}{Severity data, via Bob Verity's \code{markovid}
package. This needs to be \code{NULL} (use the default bundled data
version in the package), a \link{data.frame} object (for raw severity
data) or a list (for data that has already been processed by
\code{sircovid} for use).  New severity data comes from Bob Verity
via the markovid package, and needs to be carefully calibrated
with the progression parameters.}

\item{exp_noise}{Rate of exponential noise used in the compare
function - typically set to a large value so that noise is small
but non-zero. If set to \code{Inf} then there is no noise in the
observation process (not realistic but useful for testing).}

\item{initial_seed_size}{Initial size of seeding from the S to E
compartment; all seeding is in the 15-19 year old group from
\code{start_date} according to \code{initial_seed_pattern}. The default is 30.}

\item{initial_seed_pattern}{A vector of seeding weights for the initial
seeding. The length represents the number of steps to seed over from the
\code{start_date}, and the \code{initial_seed_size} is split over these steps
according to those weights. If \code{start_date} is not a multiple of the step
size (and thus falls between two steps) then we weight over an additional
step and adjust the weights according to how far the \code{start_date} is from
the previous full step.}
}
\value{
A list of inputs to the model, many of which are fixed and
represent data. These correspond largely to \code{user()} calls
within the odin code, though some are also used in processing
just before the model is run.
}
\description{
Parameters for the "\link{basic}" model.
}
\examples{
basic_parameters(sircovid_date("2020-02-01"), "uk")
}
