% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_compare}
\alias{carehomes_compare}
\title{Compare observed and modelled data for the carehomes model}
\usage{
carehomes_compare(state, prev_state, observed, pars)
}
\arguments{
\item{state}{State vector for the end of the current day. This is
assumed to be filtered following \code{\link[=carehomes_index]{carehomes_index()}} so contains
10 rows corresponding to ICU, general beds, admissions, deaths and
seroconversion compartments.}

\item{prev_state}{State vector for the end of the previous day, as
for \code{state}.}

\item{observed}{Observed data. This will be a list with elements
\code{icu} (number of ICU beds occupied), \code{general} (number of
general beds occupied), \code{deaths_hosp} (cumulative deaths in
hospital settings), \code{deaths_comm} (cumulative deaths in
community settings), \code{deaths} (combined deaths - used by some
nations - if given then \code{deaths_hosp} and \code{deaths_comm} must be
\code{NA}), \code{admitted} (hospital admissions), \code{new} (new cases),
\code{npos_15_64} (number of people seropositive in ages 15-64) and
\code{ntot_15_64} (number of people tested in ages 15-64), \code{pillar2_pos}
(number of pillar 2 positives), \code{pillar2_tot} (number of pillar 2
tests)}

\item{pars}{A list of parameters, as created by
\code{\link[=carehomes_parameters]{carehomes_parameters()}}}
}
\value{
A vector of log likelihoods, the same length as the number
of particles (the number of columns in the modelled state)
}
\description{
Compare observed and modelled data from the \link{carehomes} model. This
conforms to the mcstate interface.
}
