% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_forecast}
\alias{carehomes_forecast}
\title{Forecast the carehomes model}
\usage{
carehomes_forecast(
  samples,
  n_sample,
  burnin,
  forecast_days,
  incidence_states,
  prepend_trajectories = TRUE
)
}
\arguments{
\item{n_sample}{The number of samples to draw from \code{object} \emph{with
replacement}. This means that \code{n_sample} can be larger than the
total number of samples taken (though it probably should not)}

\item{burnin}{Optional integer number of iterations to discard as
"burn-in". If given then samples \code{1:burnin} will be
excluded from your results. Remember that the first sample
represents the starting point of the chain. It is an error if
this is not a positive integer or is greater than or equal to
the number of samples (i.e., there must be at least one sample
remaining after discarding burnin).}

\item{forecast_days}{The number of days to create a forecast for}

\item{incidence_states}{A character vector of states for which
incidnce should be computed (from cumulative compartments, such
as deaths). These will end up in the final trajectories object
with the sufix \verb{_inc} (e.g., \code{deaths} becomes \code{deaths_inc}).}

\item{prepend_trajectories}{Prepend trajectories from the particle
filter to the predictions created here.}
}
\description{
Forecast from the carehomes model; this provides a wrapper around
\link[mcstate:pmcmc_thin]{mcstate::pmcmc_sample} and \link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict} that samples
the trajectories then creatres samples, setting the sircovid dates
and adding trajectories of incidence.
}
