% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters}
\alias{carehomes_parameters}
\title{Parameters for the carehomes model}
\usage{
carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  p_death_carehome = 0.7,
  p_specificity = 0.9,
  progression = NULL,
  eps = 0.1,
  C_1 = 4e-06,
  C_2 = 5e-05,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99,
  prop_noncovid_sympt = 0.01,
  exp_noise = 1e+06
)
}
\arguments{
\item{start_date}{Thee start date, as a \code{\link[=sircovid_date]{sircovid_date()}} (i.e.,
the number of days into 2020)}

\item{region}{The region to run the model for. This will bee used
to get population data, which is currently fixed within the
package and is limited to "uk", the four constituent nations
("england", "wales", "scotland", "northern_ireland") and thee 7
NHS regions (e.g., "midlands"). These names are case
insensitive.}

\item{beta_date}{A vector of dates for changes in beta (the
contact rate parameter), or \code{NULL} if a single value is used for
all times (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{date}).}

\item{beta_value}{A vector of values for beta (the contact rate
parameter). If not given, and if \code{beta_date} is null then a
value of 0.08 will be used through the whole simulation,
otherwise if \code{beta_date} is \code{NULL} this must be a scalar. If
\code{beta_date} is given then \code{beta_date} and \code{beta_value} must have
the same length (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{value}).}

\item{severity}{Severity data, via Bob Verity's \code{markovid}
package. This needs to be \code{NULL} (use the default bundled data
version in the package), a \link{data.frame} object (for raw severity
data) or a list (for data that has already been processed by
\code{sircovid} for use).  New severity data comes from Bob Verity
via the markovid package, and needs to be carefully calibrated
with the progression parameters.}

\item{p_death_carehome}{Probability of death within carehomes
(conditional on having an "inflenza-like-illness")}

\item{p_specificity}{Specificity of the serology test}

\item{progression}{Progression data}

\item{eps}{Change in contact rate for carehome residents}

\item{C_1}{Contact rate between carehome workers and either
residents or workers}

\item{C_2}{Contact rate between carehome residents}

\item{pillar2_specificity}{Specificity of the Pillar 2 test}

\item{pillar2_sensitivity}{Sensitivity of the Pillar 2 test}

\item{react_specificity}{Specificity of the REACT test}

\item{react_sensitivity}{Sensitivity of the REACT test}

\item{prop_noncovid_sympt}{Proportion of population who do not have
covid but have covid-like symptoms}

\item{exp_noise}{Rate of exponential noise used in the compare
function - typically set to a large value so that noise is small
but non-zero. If set to \code{Inf} then there is no noise in the
observation process (not realistic but useful for testing).}
}
\value{
A list of inputs to the model, many of which are fixed and
represent data. These correspond largely to \code{user()} calls
within the odin code, though some are also used in processing
just before the model is run.
}
\description{
Parameters for the "\link{carehomes}" model.
}
\examples{
carehomes_parameters(sircovid_date("2020-02-01"), "uk")
}
