% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters}
\alias{carehomes_parameters}
\title{Parameters for the carehomes model}
\usage{
carehomes_parameters(
  start_date,
  region,
  beta_date = NULL,
  beta_value = NULL,
  severity = NULL,
  p_death_carehome = 0.7,
  sero_specificity = 0.9,
  sero_sensitivity = 0.99,
  progression = NULL,
  eps = 0.1,
  m_CHW = 4e-06,
  m_CHR = 5e-05,
  pillar2_specificity = 0.99,
  pillar2_sensitivity = 0.99,
  react_specificity = 0.99,
  react_sensitivity = 0.99,
  p_NC = 0.01,
  strain_transmission = 1,
  strain_seed_date = NULL,
  strain_seed_value = NULL,
  rel_susceptibility = 1,
  rel_p_sympt = 1,
  rel_p_hosp_if_sympt = 1,
  rel_infectivity = 1,
  vaccine_progression_rate = NULL,
  vaccine_schedule = NULL,
  vaccine_index_dose2 = NULL,
  waning_rate = 0,
  model_pcr_and_serology_user = 1,
  exp_noise = 1e+06
)
}
\arguments{
\item{start_date}{Thee start date, as a \code{\link[=sircovid_date]{sircovid_date()}} (i.e.,
the number of days into 2020)}

\item{region}{The region to run the model for. This will bee used
to get population data, which is currently fixed within the
package and is limited to "uk", the four constituent nations
("england", "wales", "scotland", "northern_ireland") and the 7
NHS regions (e.g., "midlands"). These names are case
insensitive.}

\item{beta_date}{A vector of dates for changes in beta (the
contact rate parameter), or \code{NULL} if a single value is used for
all times (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{date}).}

\item{beta_value}{A vector of values for beta (the contact rate
parameter). If not given, and if \code{beta_date} is \code{NULL} then a
value of 0.08 will be used through the whole simulation,
otherwise if \code{beta_date} is \code{NULL} this must be a scalar. If
\code{beta_date} is given then \code{beta_date} and \code{beta_value} must have
the same length (see \code{\link[=sircovid_parameters_beta]{sircovid_parameters_beta()}}, where this is
passed as \code{value}).}

\item{severity}{Severity data, via Bob Verity's \code{markovid}
package. This needs to be \code{NULL} (use the default bundled data
version in the package), a \link{data.frame} object (for raw severity
data) or a list (for data that has already been processed by
\code{sircovid} for use).  New severity data comes from Bob Verity
via the markovid package, and needs to be carefully calibrated
with the progression parameters.}

\item{p_death_carehome}{Probability of death within carehomes
(conditional on having an "inflenza-like-illness")}

\item{sero_specificity}{Specificity of the serology test}

\item{sero_sensitivity}{Sensitivity of the serology test}

\item{progression}{Progression data}

\item{eps}{Change in contact rate for carehome residents}

\item{m_CHW}{Contact rate between carehome workers and either
residents or workers}

\item{m_CHR}{Contact rate between carehome residents}

\item{pillar2_specificity}{Specificity of the Pillar 2 test}

\item{pillar2_sensitivity}{Sensitivity of the Pillar 2 test}

\item{react_specificity}{Specificity of the REACT test}

\item{react_sensitivity}{Sensitivity of the REACT test}

\item{p_NC}{Proportion of population who do not have
covid but have covid-like symptoms}

\item{strain_transmission}{Vector of relative transmissibility of each
strain modelled. First element should be 1. Length will define the
number of strains used in the model}

\item{strain_seed_date}{Either \code{NULL} (no seeding) or a vector of
exactly two \link{sircovid_date} values corresponding to
the start and stop dates of seeding (inclusive). For example, to
seed for a single day, these two values should have the same
date}

\item{strain_seed_value}{Either \code{NULL} (no seeding) or a single
integer value represending the \emph{daily} rate of seeding. Because
each day is dividied into \code{1 / dt} steps, this value will be
spread out fairly evenly across the steps that occur within each
date. Seeding is not stochastic; this many individuals \emph{will}
become infected with the new strain, unless the pool of
susceptibles has been exhausted.}

\item{rel_susceptibility}{A vector or matrix of values representing the
relative susceptibility of individuals in different vaccination groups.
If a vector, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1. In that case relative
susceptibility will be the same across all age groups within one
vaccination category. Specifying a matrix instead of a vector allows
different relative susceptibilities by age (rows of the matrix) and
vaccination group (columns of the matrix); in that case, in each row of
the matrix, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1}

\item{rel_p_sympt}{A vector or matrix of values of same dimension as
rel_susceptibility representing the
relative probability of symptomatic infection in different
vaccination groups. If a vector, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1.
In that case the relative reduction in probability of symptomatic
infection will be the same across all age groups within one vaccination
category.
Specifying a matrix instead of a vector allows different relative
reductions in probability of symptomatic infection by age (rows of the
matrix) and vaccination group (columns of the matrix); in that case,
in each row of the matrix, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1}

\item{rel_p_hosp_if_sympt}{A vector or matrix of values of same dimension as
rel_susceptibility representing the
relative probability of hospitalisation for symptomatic cases in different
vaccination groups. If a vector, the first value should be 1 (for the
non-vaccinated group) and subsequent values be between 0 and 1.
In that case the relative reduction in probability of hospitalisation for
symptomatic cases will be the same across all age groups within one
vaccination category.
Specifying a matrix instead of a vector allows different relative
reductions in probability of hospitalisation for symptomatic cases by age
(rows of the matrix) and vaccination group (columns of the matrix);
in that case, in each row of the matrix, the first value should be 1
(for the non-vaccinated group) and subsequent values be between 0 and 1}

\item{rel_infectivity}{A vector or matrix of values representing the
relative infectivity of individuals in different vaccination groups,
if they are infected.
If a vector, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1. In that case relative
infectivity will be the same across all age groups within one
vaccination category. Specifying a matrix instead of a vector allows
different relative infectivities by age (rows of the matrix) and
vaccination group (columns of the matrix); in that case, in each row of
the matrix, the first value should be 1 (for the non-vaccinated group)
and subsequent values be between 0 and 1}

\item{vaccine_progression_rate}{A vector or matrix of values of same
dimension as rel_susceptibility representing
the rate of movement between different vaccination classes. If a vector,
it should have as many values as vaccination classes, and the same rates
of progression will be used for all age
groups (the first rate is the vaccination rate and the last of the rates
is the rate of returning to the initial
vacination class); if a matrix, the element on row i and column j is the
rate of progression from the jth vaccination class to the (j+1)th for age
group i.}

\item{vaccine_schedule}{A \link{vaccine_schedule} object indicating the
people to be vaccinated by group over time}

\item{vaccine_index_dose2}{The index to use for the second dose}

\item{waning_rate}{A single value or a vector of values representing the
rates of waning of immunity after infection; if a single value the same
rate is used for all age groups; if a vector of values if used it should
have one value per age group.}

\item{model_pcr_and_serology_user}{A value of 1 or 0 so switch on or off the
flows out of T_PCR_neg and T_sero_neg and the corresponding cap on the
number of individuals leaving the R compartments}

\item{exp_noise}{Rate of exponential noise used in the compare
function - typically set to a large value so that noise is small
but non-zero. If set to \code{Inf} then there is no noise in the
observation process (not realistic but useful for testing).}
}
\value{
A list of inputs to the model, many of which are fixed and
represent data. These correspond largely to \code{user()} calls
within the odin code, though some are also used in processing
just before the model is run.
}
\description{
Parameters for the "\link{carehomes}" model.
}
\examples{

region <- "london"
carehomes_parameters(sircovid_date("2020-02-01"), region)

# example set up of vaccination parameters independent of age
# 3 groups: 1) unvaccinated, 2) vaccinated with partial immunity
# 3) fully vaccinated (but with an imperfect vaccine). People return
# to group 1 after a period of time in group 3 to mirror waning immunity

# Assumption: immediately after vaccination susceptibility is reduced by
# 20\%, and then by 50\% when you reach full effect of the vaccination,
# then susceptibility returns to 100\% upon waning of vaccine=induced
# immunity
# effect of vaccination similar across all age groups
rel_susceptibility <- c(1, 0.8, 0.5)

# The vaccine also reduces the risk of symptoms
rel_p_sympt <- c(1, 0.6, 0.3)
# and the risk of hospitalisation for those with symptoms
rel_p_hosp_if_sympt <- c(1, 0.95, 0.95)

# The vaccine also reduces infectivity of infected individuals by half
rel_infectivity <- c(1, 0.5, 0.5)

# On average 10000 first doses of vaccine are given every day
# Second doses are given on average 12 weeks after the first dose
# vaccine-induced immunity wanes after a period which is exponentially
# distributed and lasts on average 26 weeks (half a year);
# they are similar across all age groups
vaccine_progression_rate <- c(0, 0, 1/(26*7))

daily_doses <- rep(10000, 365) 
mean_days_between_doses <- 12 * 7
n <- vaccination_priority_population(region, uptake = 1)
schedule <- vaccination_schedule_future(
  0, daily_doses, mean_days_between_doses, n)

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2)

# vaccination parameters are automatically copied across all age groups
p$rel_susceptibility
p$rel_p_sympt
p$rel_p_hosp_if_sympt
p$rel_infectivity
# Note that this is only the "base" rate as we fill in the first
# column dynamically based on vaccine_daily_doses
p$vaccine_progression_rate

### same example as above BUT assume a different effect of vaccine in the
### first age group
n_groups <- 19

# Assumption: vaccine is twice more effective at reducing susceptibility
# in the first age group
rel_susceptibility_agegp1 <- c(1, 0.4, 0.25)
rel_susceptibility_other_agegp <- c(1, 0.8, 0.5)
rel_susceptibility <- matrix(NA, nrow = n_groups, ncol = 3)
rel_susceptibility[1, ] <- rel_susceptibility_agegp1
for (i in seq(2, n_groups)) {
  rel_susceptibility[i, ] <- rel_susceptibility_other_agegp
}
rel_susceptibility

# But vaccine has the same impact on probability of symptoms and
# hospitalisation for the symptomatic across all age groups
rel_p_sympt <- matrix(rep(rel_p_sympt, n_groups), nrow = n_groups,
  byrow = TRUE)
rel_p_hosp_if_sympt <-
  matrix(rep(rel_p_hosp_if_sympt, n_groups), nrow = n_groups, byrow = TRUE)

# And vaccine has the same impact on onwards infectivity across age groups
rel_infectivity <- matrix(rep(rel_infectivity, n_groups), nrow = n_groups,
  byrow = TRUE)

# the period of build-up of immunity is the same for all age groups,
# lasting on average 2 weeks,
# but the first age group loses immunity more quickly
# (on average after 3 months) than the other age groups
# (on average after 6 months)
vaccine_progression_rate <- cbind(0, 0,
                                  c(1 / (13 * 7),
                                  rep( 1 / (26 * 7), n_groups - 1)))

# generate model parameters
p <- carehomes_parameters(
       sircovid_date("2020-02-01"), region,
       rel_susceptibility = rel_susceptibility,
       rel_p_sympt = rel_p_sympt,
       rel_p_hosp_if_sympt = rel_p_hosp_if_sympt,
       rel_infectivity = rel_infectivity,
       vaccine_progression_rate = vaccine_progression_rate,
       vaccine_schedule = schedule,
       vaccine_index_dose2 = 2)

# TODO: add an example of manually set up vaccine schedule

}
