% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_observation}
\alias{carehomes_parameters_observation}
\title{Carehomes observation parameters}
\usage{
carehomes_parameters_observation(exp_noise = 1e+06)
}
\arguments{
\item{exp_noise}{Rate parameter for the exponentially-distributed noise used
in the likelihood calculation. Typically a large value so noise is small.}
}
\value{
A list of parameter values
}
\description{
Carehomes observation parameters
}
