% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_progression}
\alias{carehomes_parameters_progression}
\title{Carehomes progression parameters}
\usage{
carehomes_parameters_progression()
}
\value{
A list of parameter values
}
\description{
Carehomes progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
