% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_progression}
\alias{carehomes_parameters_progression}
\title{Carehomes progression parameters}
\usage{
carehomes_parameters_progression(
  dt,
  rel_gamma_A = 1,
  rel_gamma_P = 1,
  rel_gamma_C_1 = 1,
  rel_gamma_C_2 = 1,
  gamma_ICU_pre_date = NULL,
  gamma_ICU_pre_value = NULL,
  gamma_ICU_D_date = NULL,
  gamma_ICU_D_value = NULL,
  gamma_ICU_W_D_date = NULL,
  gamma_ICU_W_D_value = NULL,
  gamma_ICU_W_R_date = NULL,
  gamma_ICU_W_R_value = NULL,
  gamma_H_D_date = NULL,
  gamma_H_D_value = NULL,
  gamma_H_R_date = NULL,
  gamma_H_R_value = NULL,
  gamma_W_D_date = NULL,
  gamma_W_D_value = NULL,
  gamma_W_R_date = NULL,
  gamma_W_R_value = NULL
)
}
\arguments{
\item{rel_gamma_A}{Vector of relative rates of gamma_A for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{rel_gamma_P}{Vector of relative rates of gamma_P for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{rel_gamma_C_1}{Vector of relative rates of gamma_C_1 for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}

\item{rel_gamma_C_2}{Vector of relative rates of gamma_C_2 for each
strain modelled. If \code{1} all strains have same rates. Otherwise vector of
same length as \code{strain_transmission}, with entries that determines the
relative scaling of the defaults for each strain.}
}
\value{
A list of parameter values
}
\description{
Carehomes progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
