% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_progression}
\alias{carehomes_parameters_progression}
\title{Carehomes progression parameters}
\usage{
carehomes_parameters_progression(
  dt,
  gamma_E = list(date = NULL, value = NULL),
  gamma_A = list(date = NULL, value = NULL),
  gamma_P = list(date = NULL, value = NULL),
  gamma_C_1 = list(date = NULL, value = NULL),
  gamma_C_2 = list(date = NULL, value = NULL),
  gamma_ICU_pre = list(date = NULL, value = NULL),
  gamma_ICU_D = list(date = NULL, value = NULL),
  gamma_ICU_W_D = list(date = NULL, value = NULL),
  gamma_ICU_W_R = list(date = NULL, value = NULL),
  gamma_H_D = list(date = NULL, value = NULL),
  gamma_H_R = list(date = NULL, value = NULL),
  gamma_W_D = list(date = NULL, value = NULL),
  gamma_W_R = list(date = NULL, value = NULL),
  gamma_G_D = list(date = NULL, value = NULL)
)
}
\arguments{
\item{dt}{The step size}
}
\value{
A list of parameter values
}
\description{
Carehomes progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
