% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_progression}
\alias{carehomes_parameters_progression}
\title{Carehomes progression parameters}
\usage{
carehomes_parameters_progression(
  dt,
  gamma_E = list(date = NULL, value = NULL),
  gamma_A = list(date = NULL, value = NULL),
  gamma_P = list(date = NULL, value = NULL),
  gamma_C_1 = list(date = NULL, value = NULL),
  gamma_C_2 = list(date = NULL, value = NULL),
  gamma_ICU_pre = list(date = NULL, value = NULL),
  gamma_ICU_D = list(date = NULL, value = NULL),
  gamma_ICU_W_D = list(date = NULL, value = NULL),
  gamma_ICU_W_R = list(date = NULL, value = NULL),
  gamma_H_D = list(date = NULL, value = NULL),
  gamma_H_R = list(date = NULL, value = NULL),
  gamma_W_D = list(date = NULL, value = NULL),
  gamma_W_R = list(date = NULL, value = NULL),
  gamma_G_D = list(date = NULL, value = NULL)
)
}
\arguments{
\item{dt}{The step size}

\item{gamma_E}{A list with \code{date} and \code{value} for changes in gamma_E. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_A}{A list with \code{date} and \code{value} for changes in gamma_A. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_P}{A list with \code{date} and \code{value} for changes in gamma_P. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_C_1}{A list with \code{date} and \code{value} for changes in gamma_C_1. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_C_2}{A list with \code{date} and \code{value} for changes in gamma_C_2. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_ICU_pre}{A list with \code{date} and \code{value} for changes in
gamma_ICU_pre. If \code{value} is \code{NULL} then a default value is used. If
\code{value} is scalar then \code{date} can be \code{NULL} or unspecified. If \code{value} is
a vector then \code{date} must be a vector of sircovid dates of the same length
as \code{value}.}

\item{gamma_ICU_D}{A list with \code{date} and \code{value} for changes in
gamma_ICU_D. If \code{value} is \code{NULL} then a default value is used. If
\code{value} is scalar then \code{date} can be \code{NULL} or unspecified. If \code{value} is
a vector then \code{date} must be a vector of sircovid dates of the same length
as \code{value}.}

\item{gamma_ICU_W_D}{A list with \code{date} and \code{value} for changes in
gamma_ICU_W_D. If \code{value} is \code{NULL} then a default value is used. If
\code{value} is scalar then \code{date} can be \code{NULL} or unspecified. If \code{value} is
a vector then \code{date} must be a vector of sircovid dates of the same length
as \code{value}.}

\item{gamma_ICU_W_R}{A list with \code{date} and \code{value} for changes in
gamma_ICU_pre. If \code{value} is \code{NULL} then a default value is used. If
\code{value} is scalar then \code{date} can be \code{NULL} or unspecified. If \code{value} is
a vector then \code{date} must be a vector of sircovid dates of the same length
as \code{value}.}

\item{gamma_H_D}{A list with \code{date} and \code{value} for changes in gamma_H_D. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_H_R}{A list with \code{date} and \code{value} for changes in gamma_H_R. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_W_D}{A list with \code{date} and \code{value} for changes in gamma_W_D. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_W_R}{A list with \code{date} and \code{value} for changes in gamma_W_R. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}

\item{gamma_G_D}{A list with \code{date} and \code{value} for changes in gamma_G_D. If
\code{value} is \code{NULL} then a default value is used. If \code{value} is scalar then
\code{date} can be \code{NULL} or unspecified. If \code{value} is a vector then \code{date}
must be a vector of sircovid dates of the same length as \code{value}.}
}
\value{
A list of parameter values
}
\description{
Carehomes progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
