% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_progression}
\alias{carehomes_parameters_progression}
\title{Carehomes progression parameters}
\usage{
carehomes_parameters_progression(
  dt,
  gamma_E_date = NULL,
  gamma_E_value = NULL,
  gamma_A_date = NULL,
  gamma_A_value = NULL,
  gamma_P_date = NULL,
  gamma_P_value = NULL,
  gamma_C_1_date = NULL,
  gamma_C_1_value = NULL,
  gamma_C_2_date = NULL,
  gamma_C_2_value = NULL,
  gamma_ICU_pre_date = NULL,
  gamma_ICU_pre_value = NULL,
  gamma_ICU_D_date = NULL,
  gamma_ICU_D_value = NULL,
  gamma_ICU_W_D_date = NULL,
  gamma_ICU_W_D_value = NULL,
  gamma_ICU_W_R_date = NULL,
  gamma_ICU_W_R_value = NULL,
  gamma_H_D_date = NULL,
  gamma_H_D_value = NULL,
  gamma_H_R_date = NULL,
  gamma_H_R_value = NULL,
  gamma_W_D_date = NULL,
  gamma_W_D_value = NULL,
  gamma_W_R_date = NULL,
  gamma_W_R_value = NULL,
  gamma_G_D_date = NULL,
  gamma_G_D_value = NULL
)
}
\value{
A list of parameter values
}
\description{
Carehomes progression parameters.  The \code{s_} parameters are the
scaling parameters for the Erlang distibution (a.k.a 'k'), while
the \code{gamma_} parameters are the gamma parameters of that
distribution.  These need to be aligned with Bob's severity
outputs, and we will come up with a better way of coordinating the
two.
}
