% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes.R
\name{carehomes_parameters_severity}
\alias{carehomes_parameters_severity}
\title{Carehomes severity parameters}
\usage{
carehomes_parameters_severity(
  dt,
  severity = NULL,
  p_C_date = NULL,
  p_C_value = NULL,
  p_H_date = NULL,
  p_H_value = NULL,
  p_H_CHR_date = NULL,
  p_H_CHR_value = NULL,
  p_ICU_date = NULL,
  p_ICU_value = NULL,
  p_H_D_date = NULL,
  p_H_D_value = NULL,
  p_ICU_D_date = NULL,
  p_ICU_D_value = NULL,
  p_W_D_date = NULL,
  p_W_D_value = NULL,
  p_G_D_date = NULL,
  p_G_D_value = NULL,
  p_G_D_CHR_date = NULL,
  p_G_D_CHR_value = NULL,
  p_star_date = NULL,
  p_star_value = NULL
)
}
\value{
A list of severity parameters
}
\description{
Carehomes severity parameters
}
