% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{carehomes_Rt}
\alias{carehomes_Rt}
\title{Compute "Rt"}
\usage{
carehomes_Rt(step, S, p, type = NULL)
}
\arguments{
\item{step}{A vector of steps that the model was run over}

\item{S}{A 19 x steps matrix of "S" compartment counts}

\item{p}{A \code{\link[=carehomes_parameters]{carehomes_parameters()}} object}

\item{type}{A character vector of possible Rt types to
compute. Can be any or all of \code{eff_Rt_all}, \code{eff_Rt_general},
\code{Rt_all} and \code{Rt_general}}
}
\value{
A list with elements \code{step}, \code{beta}, and any of the \code{type}
values specified above.
}
\description{
Compute "Rt" for a single simulated trajectory and parameter set.
}
