% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{carehomes_Rt}
\alias{carehomes_Rt}
\title{Compute "Rt"}
\usage{
carehomes_Rt(step, S, p)
}
\arguments{
\item{step}{A vector of steps that the model was run over}

\item{S}{A 19 x steps matrix of "S" compartment counts}

\item{p}{A \code{\link[=carehomes_parameters]{carehomes_parameters()}} object}
}
\value{
A list with elements \code{step}, \code{beta}, \code{eff_Rt_all},
\code{eff_Rt_general}, \code{Rt_all} and \code{Rt_general}
}
\description{
Compute "Rt" for a single simulated trajectory and parameter set.
}
