% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carehomes_rt.R
\name{carehomes_Rt_trajectories}
\alias{carehomes_Rt_trajectories}
\title{Compute Rt for a set of trajectories}
\usage{
carehomes_Rt_trajectories(
  step,
  S,
  pars,
  prob_strain = NULL,
  initial_step_from_parameters = TRUE,
  shared_parameters = NULL
)
}
\arguments{
\item{step}{A vector of steps}

\item{S}{A 3d (19 x n trajectories x n steps) array of "S"
compartment counts}

\item{pars}{Either a single \code{\link[=carehomes_parameters]{carehomes_parameters()}} object
(shared parameters) or an unnamed list of
\code{\link[=carehomes_parameters]{carehomes_parameters()}} objects, the same length as \code{ncol(S)}.}

\item{prob_strain}{prob_strain outputs from model}

\item{initial_step_from_parameters}{If \code{TRUE}, then \code{step[[1]]} is
replaced by the value of \code{initial_step} from the parameters.
This is usually what you want.}

\item{shared_parameters}{Should \code{pars} be treated as a single
shared list? Leave as \code{NULL} to detect automatically, set to
\code{TRUE} or \code{FALSE} to force it to be interpreted one way or the
other which may give more easily interpretable error messages.}
}
\value{
As for \code{\link[=carehomes_Rt]{carehomes_Rt()}}, except that every element is a
matrix, not a vector.
}
\description{
Compute "Rt" for a set of simulated trajectories (e.g., the result
of \code{\link[dust:dust_iterate]{dust::dust_iterate()}}, \code{\link[mcstate:pmcmc]{mcstate::pmcmc()}} or
\code{\link[mcstate:pmcmc_predict]{mcstate::pmcmc_predict()}}. The trajectories may or may not share
parameters.
}
