% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{combine_rt}
\alias{combine_rt}
\title{Combine Rt estimates}
\usage{
combine_rt(rt, samples)
}
\arguments{
\item{rt}{A list of Rt calculations from
\code{\link[=carehomes_Rt_trajectories]{carehomes_Rt_trajectories()}} (though any Rt calculation that
confirms to this will work)}

\item{samples}{A list of samples from \code{\link[=carehomes_forecast]{carehomes_forecast()}}}
}
\value{
A list of Rt output in the same structure as the first
element of \code{rt}. All Rt estimates will be aggregated across
regions (or whatever else you are aggregating on) based on the
parameters in \code{samples}.
}
\description{
Combine Rt across multiple runs.
}
