% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{combine_trajectories}
\alias{combine_trajectories}
\title{Combine trajectories}
\usage{
combine_trajectories(samples, rank = TRUE)
}
\arguments{
\item{samples}{A list of samples from \code{\link[=lancelot_forecast]{lancelot_forecast()}}}

\item{rank}{Logical, indicating if trajectories should be ranked
before combination.}
}
\value{
A set of combined trajectories (not a combined samples object).
}
\description{
Combine trajectories across multiple runs
}
