% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_output}
\alias{compare_output}
\title{Compare model to ICU data}
\usage{
compare_output(model, pars_obs, data, type = "sircovid_basic")
}
\arguments{
\item{model}{An \code{odin_model} object}

\item{pars_obs}{Parameters for the observations}

\item{data}{The data to be compared against}

\item{type}{The class of the model, either
\code{"sircovid_basic"} or \code{"sircovid_hospital"}}
}
\description{
Compare the model to ICU data for use with the particle filter
}
