% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_beta}
\alias{generate_beta}
\title{Generate beta}
\usage{
generate_beta(
  beta_start,
  start_date = "2020-02-02",
  reduction_start = "2020-03-16",
  beta_reduction = 0.238,
  reduction_period = 10
)
}
\arguments{
\item{beta_start}{Value of beta before intervention}

\item{start_date}{Start date for simualtions}

\item{reduction_start}{Start of reduction in beta}

\item{beta_reduction}{Factor by which beta is reduced
default is 0.238, as estimated by Jarvis et al
beta_reduced = beta - beta * (1 - beta_reduction)}

\item{reduction_period}{Time, in days, over which the
reduction in beta is achieved. Default 10 days}
}
\description{
Create a linear time-varying beta for input into 
\code{create_parameters()}
}
