% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\alias{generate_parameters_new_hospital_model}
\title{Create parameters}
\usage{
generate_parameters(
  transmission_model = "POLYMOD",
  country = "United Kingdom",
  severity_data_file = NULL,
  progression_groups = list(E = 2, asympt = 1, mild = 1, ILI = 1, hosp = 2, ICU = 2, rec
    = 2),
  gammas = list(E = 1/(4.59/2), asympt = 1/2.09, mild = 1/2.09, ILI = 1/4, hosp = 2, ICU
    = 2/5, rec = 2/5),
  infection_seeding = list(values = c(10), bins = c("15 to 19")),
  beta = c(0.1, 0.1, 0.1),
  beta_times = c("2020-02-02", "2020-03-01", "2020-04-01"),
  trans_profile = c(1),
  trans_increase = c(1),
  hosp_transmission = 0.1,
  ICU_transmission = 0.05,
  dt = 0.25,
  use_polymod_pop = FALSE
)

generate_parameters_new_hospital_model()
}
\arguments{
\item{transmission_model}{Model type. Only 'POLYMOD' currently supported}

\item{country}{Country name}

\item{severity_data_file}{Location of file with severity data}

\item{progression_groups}{List of number of progression groups in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}

\item{gammas}{List of exponential distribution rates for time in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}

\item{infection_seeding}{List of vector \code{values} of how many cases to seed 
in a correspond vector of age bins in \code{bins}.}

\item{beta}{Beta, for each time step in \code{beta_times}}

\item{beta_times}{Dates \code{beta} changes, in format yyyy-mm-dd}

\item{trans_profile}{Proportion in each infectivity group}

\item{trans_increase}{Relative infectivity of each group}

\item{hosp_transmission}{Transmission rate of hospital cases}

\item{ICU_transmission}{Tranmissions rate of ICU cases}

\item{dt}{Time-step to run the model in days}

\item{use_polymod_pop}{Set to ignore \code{survey_pop_in}
and use the population from polymod when estimating the
transmission matrix}
}
\value{
List of parameters for use with \code{sircovid}
}
\description{
Create parameters for use with the model
}
