% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Create parameters}
\usage{
generate_parameters(
  transmission_model = "POLYMOD",
  country = "United Kingdom",
  severity_data_file = "extdata/Final_COVID_severity.csv",
  age_limits = c(0, 10, 20, 30, 40, 50, 60, 70, 80),
  progression_groups = list(E = 2, asympt = 2, mild = 2, ILI = 2, hosp = 2, ICU = 2, rec
    = 2),
  gammas = list(E = 1/(4.59/2), asympt = 1, mild = 1, ILI = 1, hosp = 2/7, ICU = 2/7,
    rec = 2/3),
  infection_seeding = c(0, 0, 0, 10, 0, 0, 0, 0, 0),
  beta = c(0.1, 0.1, 0.1),
  beta_times = c("2020-02-02", "2020-03-01", "2020-04-01"),
  trans_profile = c(0.65, 0.2, 0.15),
  trans_increase = c(0, 1, 10),
  hosp_transmission = 0.1,
  ICU_transmission = 0.05,
  dt = 0.25,
  survey_pop_in = NULL
)
}
\arguments{
\item{transmission_model}{Model type. Only 'POLYMOD' currently supported}

\item{country}{Country name}

\item{severity_data_file}{Location of file with severity data}

\item{age_limits}{Vector of age bin (starting age of each bin)}

\item{progression_groups}{List of number of progression groups in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}

\item{gammas}{List of exponential distribution rates for time in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp', 'ICU', 'rec'}

\item{infection_seeding}{Vector of initial cases in each age bin. These are seeded
into the most infective group}

\item{beta}{Beta, for each time step in \code{beta_times}}

\item{beta_times}{Dates \code{beta} changes, in format yyyy-mm-dd}

\item{trans_profile}{Proportion in each infectivity group}

\item{trans_increase}{Relative infectivity of each group}

\item{hosp_transmission}{Transmission rate of hospital cases}

\item{ICU_transmission}{Tranmissions rate of ICU cases}

\item{dt}{Time-step to run the model in days}

\item{survey_pop_in}{A survey population. Set to NULL to
use \code{default_age_distribution()}}
}
\description{
Create parameters for use with the model
}
