% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{generate_parameters}
\alias{generate_parameters}
\title{Create parameters}
\usage{
generate_parameters(
  sircovid_model = basic_model(),
  transmission_model = "POLYMOD",
  country = "United Kingdom",
  severity_data_file = NULL,
  infection_seeding = list(values = c(10), bins = c("15 to 19")),
  beta = 0.1,
  beta_times = sircovid_date("2020-02-02"),
  importation = FALSE,
  importation_levels = 100,
  importation_times = sircovid_date("2020-02-02"),
  psi = 0,
  trans_profile = c(1),
  trans_increase = c(1),
  hosp_transmission = 0.1,
  ICU_transmission = 0.05,
  comm_D_transmission = 0.05,
  dt = 0.25,
  use_polymod_pop = FALSE
)
}
\arguments{
\item{sircovid_model}{Model to use, \code{basic_model()} or
\code{hospital_model()}}

\item{transmission_model}{Model type. Only 'POLYMOD' currently supported}

\item{country}{Country name}

\item{severity_data_file}{Location of file with severity data}

\item{infection_seeding}{List of vector \code{values} of how many cases to seed 
in a correspond vector of age bins in \code{bins}.}

\item{beta}{Beta, for each time step in \code{beta_times}}

\item{beta_times}{Dates \code{beta} changes, must use \code{sircovid_date}}

\item{importation}{Logical, if TRUE then there is importation in the model and there are zero initial
infectives}

\item{importation_levels}{importation levels, for each time step in \code{importation_times}}

\item{importation_times}{Dates importation changes, must use \code{sircovid_date}}

\item{psi}{parameter for scaling importation}

\item{trans_profile}{Proportion in each infectivity group}

\item{trans_increase}{Relative infectivity of each group}

\item{hosp_transmission}{Transmission rate of hospital cases}

\item{ICU_transmission}{Tranmissions rate of ICU cases}

\item{comm_D_transmission}{Transmission rate of cases dying in the community}

\item{dt}{Time-step to run the model in days}

\item{use_polymod_pop}{Set to ignore \code{survey_pop_in}
and use the population from polymod when estimating the
transmission matrix}
}
\value{
List of parameters for use with \code{sircovid}
}
\description{
Create parameters for use with the model
}
