% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{modify_severity}
\alias{modify_severity}
\title{Modify severity and transmission of VOC}
\usage{
modify_severity(efficacy, efficacy_strain_2, strain_severity_modifier)
}
\arguments{
\item{efficacy, efficacy_strain_2}{Vaccine efficacy parameters for strains
1 and 2 respectively. Expects a list
with names rel_susceptibility, rel_p_sympt, rel_p_hosp_if_sympt,
rel_infectivity, rel_p_death. Columns correspond to vaccine strata and
rows to age groups.
'}
}
\description{
Applies a modifier to severity and transmission parameters
}
