% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{reorder_rt_ifr}
\alias{reorder_rt_ifr}
\title{Reorder Rt or IFR trajectories}
\usage{
reorder_rt_ifr(x, rank)
}
\arguments{
\item{x}{An \code{Rt_trajectories} or \code{IFR_t_trajectories} object, as returned
by \code{\link[=lancelot_Rt_trajectories]{lancelot_Rt_trajectories()}} or \link{lancelot_ifr_t_trajectories}
respectively.}

\item{rank}{A vector of ranks to reorder by}
}
\value{
An \code{Rt_trajectories} or \code{IFR_t_trajectories} object with
appropriately reordered elements
}
\description{
Reorder Rt or IFR trajectories
}
