% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rotate_strains}
\alias{rotate_strains}
\title{Rotate strains}
\usage{
rotate_strains(state, info)
}
\arguments{
\item{state}{Model state}

\item{info}{Model info}
}
\description{
Rotate strains, so that strain 1 becomes the sum of strains 1 and
2 and strain 2 is empty. Use this to allow sequential replacement
of strains.
}
