% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sircovid.R
\name{run_particle_filter}
\alias{run_particle_filter}
\title{Run particle filter}
\usage{
run_particle_filter(
  data,
  model_params,
  model_start_date = "2020-02-02",
  obs_params = list(phi_ICU = 0.95, k_ICU = 2, phi_death = 1789/1651, k_death = 2,
    exp_noise = 1e+06),
  n_particles = 1000,
  forecast_days = 0,
  save_particles = FALSE,
  return = "full"
)
}
\arguments{
\item{data}{to fit to, generated by 
\code{data_from_deaths_admissions}}

\item{model_params}{Parameters to run model simulations,
generated by \code{generate_parameters()}}

\item{model_start_date}{Date to run model simulations from}

\item{obs_params}{List of parameters used for comparing
model to data in the particle filter}

\item{n_particles}{Number of particles}

\item{forecast_days}{Days ahead to include in output}

\item{save_particles}{Whether to save trajectories}

\item{return}{Set return depending on what is needed. 'full' gives
the entire particle filter output, 'll' gives the
log-likelihood, 'sample' gives a sampled particle's
trace}
}
\value{
Results from particle filter
}
\description{
Create a model, and fit with the particle filter
}
