% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid.R
\name{sample_grid_scan}
\alias{sample_grid_scan}
\title{Sample Grid Scan}
\usage{
sample_grid_scan(scan_results, n_sample_pairs = 10, n_particles = 100)
}
\arguments{
\item{scan_results}{Output of \code{\link{scan_beta_date}}.}

\item{n_sample_pairs}{Number of parameter pairs to be sampled. Integer.
Default = 10.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}
}
\value{
List. First element is an array of trajectories, in which
 first dimension is time, second is parameters chosen when sampling from 
 the grid and the third dimension is alist of inputs.
}
\description{
Take a grid search produced by \code{\link{scan_beta_date}} and 
sample \code{n_sample_pairs} from the parameter grid uses based
on their probaility. For each parameter pair chosen, run particle
filter with \code{num_particles} and sample 1 trajectory
}
