% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{scan_beta_date}
\alias{scan_beta_date}
\title{Grid search of beta and start date}
\usage{
scan_beta_date(
  min_beta,
  max_beta,
  beta_step,
  generate_beta_func = function(beta_start, start_date) {     generate_beta(beta_start =
    beta_start, start_date = start_date) },
  first_start_date,
  last_start_date,
  day_step,
  data,
  model = NULL,
  model_params = NULL,
  pars_obs = NULL,
  n_particles = 100
)
}
\arguments{
\item{min_beta}{Minimum value of beta in the search}

\item{max_beta}{Maximum value of beta in the search}

\item{beta_step}{Step to increment beta between min and max}

\item{generate_beta_func}{Function which gives a time series
of beta, given an initial value and start date. Default
is \code{generate_beta()}}

\item{first_start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{last_start_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{data}{Hosptial data to fit to. See \code{example.csv}
and \code{particle_filter_data()}}

\item{model}{An odin model generator and comparison function.  If
\code{NULL} uses the basic model.}

\item{model_params}{Model parameters, from a call to
\code{generate_parameters()}. If NULL, uses defaults as
in unit tests.}

\item{pars_obs}{list of parameters to use for the comparison function.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over beta and start date.
This is parallelised, first run \code{plan(multiprocess)} to set 
this up.
}
