% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{scan_beta_date}
\alias{scan_beta_date}
\title{Grid search of beta and start date}
\usage{
scan_beta_date(
  min_beta,
  max_beta,
  beta_step,
  start_date,
  end_date,
  day_step,
  data,
  model_params = NULL,
  pars_obs = list(phi_ICU = 0.95, k_ICU = 2, phi_death = 926/1019, k_death = 2,
    exp_noise = 1e+06)
)
}
\arguments{
\item{min_beta}{Minimum value of beta in the search}

\item{max_beta}{Maximum value of beta in the search}

\item{beta_step}{Step to increment beta between min and max}

\item{start_date}{Earliest start date as 'yyyy-mm-dd'}

\item{end_date}{Latest start date as 'yyyy-mm-dd'}

\item{day_step}{Step to increment date in days}

\item{data}{Hosptial data to fit to. See \code{example.csv}
and \code{particle_filter_data()}}

\item{model_params}{Model parameters, from a call to
\code{generate_parameters()}. If NULL, uses defaults as
in unit tests.}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over beta and start date
}
