% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeding.R
\name{seeding_function}
\alias{seeding_function}
\title{Compare model to ICU data}
\usage{
seeding_function(model, data, pars_seeding = NULL)
}
\arguments{
\item{model}{An \code{odin_model} object}

\item{data}{The data, must be constructed with
\code{particle_filter_data}}

\item{pars_seeding}{A list of parameters used for seeding,
e.g. pars_seeding = list(lambda = 20)}
}
\description{
Seeding function for use with the particle filter
}
