% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{serology2_model}
\alias{serology2_model}
\title{serology2 model}
\usage{
serology2_model(
  use_fitted_parameters = TRUE,
  progression_groups = list(E = 2, asympt = 1, mild = 1, ILI = 1, comm_D = 2, hosp_D =
    2, hosp_R = 2, ICU_D = 2, ICU_R = 2, triage = 2, stepdown = 2, PCR_pos = 2),
  gammas = list(E = 1/(4.59/2), asympt = 1/2.09, mild = 1/2.09, ILI = 1/4, comm_D =
    2/5, hosp_D = 2/5, hosp_R = 2/10, ICU_D = 2/5, ICU_R = 2/10, triage = 2, stepdown =
    2/5, R_pre_1 = 1/5, R_pre_2 = 1/10, test = 3/10, PCR_pos = 1/5)
)
}
\arguments{
\item{use_fitted_parameters}{Override progression_groups and gammas with fitted
parameters loaded by \code{read_fitted_parameters()}}

\item{progression_groups}{List of number of progression groups in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp_D', 'hosp_R', 'ICU_D', 'ICU_R', 'triage', 'stepdown', 'PCR_pos'}

\item{gammas}{List of exponential distribution rates for time in each partition
needs 'E', 'asympt', 'mild', 'ILI', 'hosp_D', 'hosp_R', 'ICU_D', 'ICU_R', 'triage', 'stepdown', 'R_pre_1', 'R_pre_2', 'test', 'PCR_pos'}
}
\description{
Create a serology2 model
}
