% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_grid.R
\name{summary.sample_grid_search}
\alias{summary.sample_grid_search}
\title{Summarise grid forecast}
\usage{
\method{summary}{sample_grid_search}(
  x,
  ...,
  what = "deaths",
  quantiles = seq(from = 0.05, to = 0.95, by = 0.05)
)
}
\arguments{
\item{what}{}

\item{quantiles}{Quantiles to summarise forecast variance.}

\item{sample_grid_results}{Results from \code{\link{sample_grid_scan}}}
}
\description{
Take a sampled grid search produced by \code{\link{sample_grid_scan}} and 
extract forecasts and quantiles in text format
}
