% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_future}
\alias{vaccine_schedule_future}
\title{Create vaccination schedule}
\usage{
vaccine_schedule_future(
  start,
  daily_doses_value,
  mean_days_between_doses,
  priority_population
)
}
\arguments{
\item{start}{Either a \link{sircovid_date} object corresponding to the
first date in daily_doses_value, or a \link{vaccine_schedule} object
corresponding to previously carried out vaccination.}

\item{daily_doses_value}{A vector of doses per day}

\item{mean_days_between_doses}{Assumed mean days between doses one
and two}

\item{priority_population}{Output from
\link{vaccine_priority_population}, giving the number of people
to vaccinate in each age (row) and priority group (column)}
}
\description{
Create future vaccination schedule from a projected number of
daily doses.
}
