% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccination.R
\name{vaccine_schedule_scenario}
\alias{vaccine_schedule_scenario}
\title{High-level vaccine scenario creation}
\usage{
vaccine_schedule_scenario(
  schedule_past,
  doses_future,
  end_date,
  mean_days_between_doses,
  priority_population,
  lag_groups = NULL,
  lag_days = NULL,
  boosters_future = NULL
)
}
\arguments{
\item{schedule_past}{A \link{vaccine_schedule} object corresponding to
previously carried out vaccination.}

\item{doses_future}{A named vector of vaccine doses to give in the
future. Names must be in ISO date format.}

\item{end_date}{The final day in the future to create a schedule
for. After this date the model will assume 0 vaccine doses given
so an overestimate is probably better than an underestimate.}

\item{mean_days_between_doses}{Assumed mean days between doses one
and two}

\item{priority_population}{Output from
\link{vaccine_priority_population}, giving the number of people
to vaccinate in each age (row) and priority group (column)}

\item{lag_groups}{Row indices, corresponding to age
groups in which a lag should be added to the start time of the dose
schedule returned by \link{vaccine_schedule}, if NULL then no lag is added.
Ignored if \code{lag_groups} is NULL.}

\item{lag_days}{If \code{lag_groups} is not NULL then specifies the number of
days to add the start of the dose schedule for the given groups. Ignored
if \code{lag_groups} is NULL.}

\item{boosters_future}{Optional named vector of booster doses to give in
the future. Names must be in ISO date format.}
}
\value{
A \link{vaccine_schedule} object
}
\description{
Create a vaccination scenario
}
